% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-security_scheme.R
\name{abstract_security_scheme}
\alias{abstract_security_scheme}
\title{Security scheme objects}
\usage{
abstract_security_scheme()
}
\description{
This is an abstract class that is used to define specific types of security
schemes.
}
\seealso{
\code{\link[=class_api_key_security_scheme]{class_api_key_security_scheme()}}, \code{\link[=class_oauth2_security_scheme]{class_oauth2_security_scheme()}}
}
\keyword{internal}
