% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-01-contact.R
\name{api_contact}
\alias{api_contact}
\title{Contact information for the API}
\usage{
api_contact(
  name = S7::class_missing,
  email = S7::class_missing,
  url = S7::class_missing,
  ...
)
}
\arguments{
\item{name}{The identifying name of the contact person/organization.}

\item{email}{The email address of the contact person/organization. This
\emph{must} be in the form of an email address.}

\item{url}{The URL pointing to the contact information. This \emph{must} be in the
form of a URL.}

\item{...}{Catch-all for unimplemented properties.}
}
\value{
An \code{api_contact} S7 object, with fields \code{name}, \code{email}, and \code{url}.
}
\description{
Validate the contact information for an API.
}
\examples{
api_contact(
  "API Support",
  "support@example.com",
  "https://www.example.com/support"
)
}
