% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-license.R
\name{api_license}
\alias{api_license}
\title{License  information for the API}
\usage{
api_license(name, ..., identifier = NULL, url = NULL)
}
\arguments{
\item{name}{The license name used for the API.}

\item{...}{These dots are for future extensions and must be empty.}

\item{identifier}{An
\href{https://spdx.org/spdx-specification-21-web-version#h.jxpfx0ykyb60}{SPDX}
license expression for the API. The \code{identifier} field is mutually
exclusive of the \code{url} field.}

\item{url}{A URL to the license used for the API. This \emph{must} be in the form
of a URL. The \code{url} field is mutually exclusive of the \code{identifier} field.}
}
\value{
A \code{rapid_license} object: a list with fields \code{name}, identifier, and
\code{url}.
}
\description{
Validate the license information for an API.
}
\examples{
api_license(
  "Apache 2.0",
  identifier = "Apache-2.0"
)
api_license(
  "Apache 2.0",
  url = "https://opensource.org/license/apache-2-0/"
)
}
