% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-01-contact.R
\name{as_api_contact}
\alias{as_api_contact}
\title{Coerce lists and character vectors to api_contacts}
\usage{
as_api_contact(x, ...)
}
\arguments{
\item{x}{The object to coerce. Must be empty or have names "name", "email",
and/or "url". Extra names are ignored.}
}
\value{
An \code{api_contact} as returned by \code{\link[=api_contact]{api_contact()}}.
}
\description{
\code{as_api_contact()} turns an existing object into an \code{api_contact}. This is in
contrast with \code{\link[=api_contact]{api_contact()}}, which builds an \code{api_contact} from individual
properties.
}
\examples{
as_api_contact()
as_api_contact(list(name = "Jon Harmon", email = "jonthegeek@gmail.com"))
}
