% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-security_scheme-api_key.R
\name{as_api_key_security_scheme}
\alias{as_api_key_security_scheme}
\title{Coerce lists and character vectors to API key security schemes}
\usage{
as_api_key_security_scheme(x, ..., arg = caller_arg(x), call = caller_env())
}
\arguments{
\item{x}{The object to coerce. Must be empty or be a list or character vector
with names "name" and either "in" or "location", or names that can be
coerced to those names via \code{\link[snakecase:to_snake_case]{snakecase::to_snake_case()}}. Additional names
are ignored.}

\item{...}{These dots are for future extensions and must be empty.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
An \code{api_key_security_scheme} as returned by
\code{\link[=class_api_key_security_scheme]{class_api_key_security_scheme()}}.
}
\description{
\code{as_api_key_security_scheme()} turns an existing object into an
\code{api_key_security_scheme}. This is in contrast with
\code{\link[=class_api_key_security_scheme]{class_api_key_security_scheme()}}, which builds an \code{api_key_security_scheme}
from individual properties.
}
\seealso{
security scheme objects:
\code{\link[=as_oauth2_authorization_code_flow]{as_oauth2_authorization_code_flow()}},
\code{\link[=as_oauth2_implicit_flow]{as_oauth2_implicit_flow()}},
\code{\link[=as_oauth2_security_scheme]{as_oauth2_security_scheme()}},
\code{\link[=as_oauth2_token_flow]{as_oauth2_token_flow()}},
\code{\link[=as_scopes]{as_scopes()}},
\code{\link[=as_security_scheme]{as_security_scheme()}},
\code{\link[=as_security_scheme_details]{as_security_scheme_details()}},
\code{\link[=as_security_schemes]{as_security_schemes()}},
\code{\link[=class_api_key_security_scheme]{class_api_key_security_scheme()}},
\code{\link[=class_oauth2_authorization_code_flow]{class_oauth2_authorization_code_flow()}},
\code{\link[=class_oauth2_implicit_flow]{class_oauth2_implicit_flow()}},
\code{\link[=class_oauth2_security_scheme]{class_oauth2_security_scheme()}},
\code{\link[=class_oauth2_token_flow]{class_oauth2_token_flow()}},
\code{\link[=class_scopes]{class_scopes()}},
\code{\link[=class_security_scheme_details]{class_security_scheme_details()}},
\code{\link[=class_security_schemes]{class_security_schemes()}}

components object:
\code{\link[=as_components]{as_components()}},
\code{\link[=as_oauth2_authorization_code_flow]{as_oauth2_authorization_code_flow()}},
\code{\link[=as_oauth2_implicit_flow]{as_oauth2_implicit_flow()}},
\code{\link[=as_oauth2_security_scheme]{as_oauth2_security_scheme()}},
\code{\link[=as_oauth2_token_flow]{as_oauth2_token_flow()}},
\code{\link[=as_reference]{as_reference()}},
\code{\link[=as_schema]{as_schema()}},
\code{\link[=as_scopes]{as_scopes()}},
\code{\link[=as_security_scheme]{as_security_scheme()}},
\code{\link[=as_security_scheme_details]{as_security_scheme_details()}},
\code{\link[=as_security_schemes]{as_security_schemes()}},
\code{\link[=class_api_key_security_scheme]{class_api_key_security_scheme()}},
\code{\link[=class_components]{class_components()}},
\code{\link[=class_oauth2_authorization_code_flow]{class_oauth2_authorization_code_flow()}},
\code{\link[=class_oauth2_implicit_flow]{class_oauth2_implicit_flow()}},
\code{\link[=class_oauth2_security_scheme]{class_oauth2_security_scheme()}},
\code{\link[=class_oauth2_token_flow]{class_oauth2_token_flow()}},
\code{\link[=class_reference]{class_reference()}},
\code{\link[=class_schema]{class_schema()}},
\code{\link[=class_scopes]{class_scopes()}},
\code{\link[=class_security_scheme_details]{class_security_scheme_details()}},
\code{\link[=class_security_schemes]{class_security_schemes()}}
}
\concept{components}
\concept{components_security_schemes}
