% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_list.R
\name{as_list}
\alias{as_list}
\title{Coerce rapid objects to lists}
\usage{
as_list(x, ...)
}
\arguments{
\item{x}{The object to coerce. Currently supports conversion of
\code{\link[=security_scheme_collection]{security_scheme_collection()}} objects.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A list.
}
\description{
Convert \code{rapid} objects to lists.
}
