% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-origin.R
\name{as_origin}
\alias{as_origin}
\title{Coerce lists and character vectors to origin}
\usage{
as_origin(x, ..., arg = caller_arg(x), call = caller_env())
}
\arguments{
\item{x}{The object to coerce. Must be empty or have names "url", "format",
and/or "version", or names that can be coerced to those names via
\code{\link[snakecase:to_snake_case]{snakecase::to_snake_case()}}. Extra names are ignored. This object should
describe a single origin for this API description.}

\item{...}{These dots are for future extensions and must be empty.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
An \code{origin} as returned by \code{\link[=class_origin]{class_origin()}}.
}
\description{
\code{as_origin()} turns an existing object into an \code{origin}. This is in contrast
with \code{\link[=class_origin]{class_origin()}}, which builds an \code{origin} from individual properties.
}
\examples{
as_origin()
as_origin(
  list(
    list(
      format = "openapi",
      url = "https://api.open.fec.gov/swagger/",
      version = "3.0"
    )
  )
)
}
\seealso{
info object:
\code{\link[=as_contact]{as_contact()}},
\code{\link[=as_info]{as_info()}},
\code{\link[=as_license]{as_license()}},
\code{\link[=class_contact]{class_contact()}},
\code{\link[=class_info]{class_info()}},
\code{\link[=class_license]{class_license()}},
\code{\link[=class_origin]{class_origin()}}
}
\concept{info}
