% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-security_scheme-oauth2-scopes.R
\name{as_scopes}
\alias{as_scopes}
\title{Coerce lists and character vectors to scopes}
\usage{
as_scopes(x, ...)
}
\arguments{
\item{x}{The object to coerce. Must be coercible to a named character vector.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A \code{scopes} as returned by \code{\link[=class_scopes]{class_scopes()}}.
}
\description{
\code{as_scopes()} turns an existing object into a \code{scopes}. This is in contrast
with \code{\link[=class_scopes]{class_scopes()}}, which builds a \code{scopes} from individual properties.
}
\seealso{
security scheme objects:
\code{\link[=as_api_key_security_scheme]{as_api_key_security_scheme()}},
\code{\link[=as_oauth2_authorization_code_flow]{as_oauth2_authorization_code_flow()}},
\code{\link[=as_oauth2_implicit_flow]{as_oauth2_implicit_flow()}},
\code{\link[=as_oauth2_security_scheme]{as_oauth2_security_scheme()}},
\code{\link[=as_oauth2_token_flow]{as_oauth2_token_flow()}},
\code{\link[=as_security_scheme]{as_security_scheme()}},
\code{\link[=as_security_scheme_details]{as_security_scheme_details()}},
\code{\link[=as_security_schemes]{as_security_schemes()}},
\code{\link[=class_api_key_security_scheme]{class_api_key_security_scheme()}},
\code{\link[=class_oauth2_authorization_code_flow]{class_oauth2_authorization_code_flow()}},
\code{\link[=class_oauth2_implicit_flow]{class_oauth2_implicit_flow()}},
\code{\link[=class_oauth2_security_scheme]{class_oauth2_security_scheme()}},
\code{\link[=class_oauth2_token_flow]{class_oauth2_token_flow()}},
\code{\link[=class_scopes]{class_scopes()}},
\code{\link[=class_security_scheme_details]{class_security_scheme_details()}},
\code{\link[=class_security_schemes]{class_security_schemes()}}

components object:
\code{\link[=as_api_key_security_scheme]{as_api_key_security_scheme()}},
\code{\link[=as_components]{as_components()}},
\code{\link[=as_oauth2_authorization_code_flow]{as_oauth2_authorization_code_flow()}},
\code{\link[=as_oauth2_implicit_flow]{as_oauth2_implicit_flow()}},
\code{\link[=as_oauth2_security_scheme]{as_oauth2_security_scheme()}},
\code{\link[=as_oauth2_token_flow]{as_oauth2_token_flow()}},
\code{\link[=as_reference]{as_reference()}},
\code{\link[=as_schema]{as_schema()}},
\code{\link[=as_security_scheme]{as_security_scheme()}},
\code{\link[=as_security_scheme_details]{as_security_scheme_details()}},
\code{\link[=as_security_schemes]{as_security_schemes()}},
\code{\link[=class_api_key_security_scheme]{class_api_key_security_scheme()}},
\code{\link[=class_components]{class_components()}},
\code{\link[=class_oauth2_authorization_code_flow]{class_oauth2_authorization_code_flow()}},
\code{\link[=class_oauth2_implicit_flow]{class_oauth2_implicit_flow()}},
\code{\link[=class_oauth2_security_scheme]{class_oauth2_security_scheme()}},
\code{\link[=class_oauth2_token_flow]{class_oauth2_token_flow()}},
\code{\link[=class_reference]{class_reference()}},
\code{\link[=class_schema]{class_schema()}},
\code{\link[=class_scopes]{class_scopes()}},
\code{\link[=class_security_scheme_details]{class_security_scheme_details()}},
\code{\link[=class_security_schemes]{class_security_schemes()}}
}
\concept{components}
\concept{components_security_schemes}
