% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-security_scheme-oauth2-scopes_list.R
\name{as_scopes_list}
\alias{as_scopes_list}
\title{Coerce lists and character vectors to scopes_lists}
\usage{
as_scopes_list(x, ...)
}
\arguments{
\item{x}{The object to coerce. Must be empty or be a list of objects that can
be coerced to \code{scopes} objects via \code{\link[=as_scopes]{as_scopes()}}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A \code{scopes_list} object as returned by \code{\link[=scopes_list]{scopes_list()}}.
}
\description{
\code{as_scopes_list()} turns an existing object into a \code{scopes_list}
object. This is in contrast with \code{\link[=scopes_list]{scopes_list()}}, which builds a
\code{scopes_list} object from individual properties.
}
