% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security_requirements.R
\name{as_security_requirements}
\alias{as_security_requirements}
\title{Coerce lists to as_security_requirements objects}
\usage{
as_security_requirements(x, ...)
}
\arguments{
\item{x}{The object to coerce. Must be empty or be a list containing a single
list named "security_schemes", or a name that can be coerced to
"security_schemes" via \code{\link[snakecase:caseconverter]{snakecase::to_snake_case()}}. Additional names are
ignored.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A \code{security_requirements} object as returned by
\code{\link[=security_requirements]{security_requirements()}}.
}
\description{
\code{as_security_requirements()} turns an existing object into a
\code{security_requirements} object. This is in contrast with
\code{\link[=security_requirements]{security_requirements()}}, which builds a \code{security_requirements} from
individual properties.
}
\examples{
as_security_requirements()
as_security_requirements(
  list(
    list(
      oauth2 = c("user", "user:email", "user:follow")
    ),
    list(internalApiKey = list())
  )
)
}
