% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers-01-server_variable.R
\name{as_server_variable}
\alias{as_server_variable}
\title{Coerce lists and character vectors to server_variables}
\usage{
as_server_variable(x, ...)
}
\arguments{
\item{x}{The object to coerce. Must be empty or have names "name", "email",
and/or "url". Extra names are ignored.}
}
\value{
A \code{server_variable} as returned by \code{\link[=server_variable]{server_variable()}}.
}
\description{
\code{as_server_variable()} turns an existing object into a \code{server_variable}.
This is in contrast with \code{\link[=server_variable]{server_variable()}}, which builds a
\code{server_variable} from individual properties.
}
\examples{
as_server_variable()
as_server_variable(list(name = "Jon Harmon", email = "jonthegeek@gmail.com"))
}
