% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers.R
\name{as_servers}
\alias{as_servers}
\title{Coerce lists and character vectors to servers}
\usage{
as_servers(x, ...)
}
\arguments{
\item{x}{The object to coerce. Must be empty or have names "name", "email",
and/or "url", or names that can be coerced to those names via
\code{\link[snakecase:to_snake_case]{snakecase::to_snake_case()}}. Extra names are ignored.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A \code{servers} object as returned by \code{\link[=class_servers]{class_servers()}}.
}
\description{
\code{as_servers()} turns an existing object into a \code{servers} object. This is in
contrast with \code{\link[=class_servers]{class_servers()}}, which builds a \code{servers} object from
individual properties.
}
\examples{
as_servers()
as_servers(
  list(
    list(
      url = "https://development.gigantic-server.com/v1",
      description = "Development server"
    ),
    list(
      url = "https://staging.gigantic-server.com/v1",
      description = "Staging server"
    ),
    list(
      url = "https://api.gigantic-server.com/v1",
      description = "Production server"
    )
  )
)
}
\seealso{
servers object:
\code{\link[=as_server_variables]{as_server_variables()}},
\code{\link[=as_string_replacements]{as_string_replacements()}},
\code{\link[=class_server_variables]{class_server_variables()}},
\code{\link[=class_servers]{class_servers()}},
\code{\link[=class_string_replacements]{class_string_replacements()}}
}
\concept{servers}
