% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers-02-variables.R
\name{as_variables}
\alias{as_variables}
\title{Coerce lists and character vectors to variables}
\usage{
as_variables(x, ...)
}
\arguments{
\item{x}{The object to coerce. Must be empty or have names "name", "email",
and/or "url". Extra names are ignored.}
}
\value{
A \code{variables} as returned by \code{\link[=variables]{variables()}}.
}
\description{
\code{as_variables()} turns an existing object into a \code{variables}. This is in
contrast with \code{\link[=variables]{variables()}}, which builds a \code{variables} from individual
properties.
}
\examples{
as_variables()
as_variables(list(name = "Jon Harmon", email = "jonthegeek@gmail.com"))
}
