% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-contact.R
\name{class_contact}
\alias{class_contact}
\title{Contact information for the API}
\usage{
class_contact(name = character(0), email = character(0), url = character(0))
}
\arguments{
\item{name}{Character scalar (optional). The identifying name of the contact
person or organization.}

\item{email}{Character scalar (optional). The email address of the contact
person/organization. This must be in the form of an email address.}

\item{url}{Character scalar (optional). The URL pointing to the contact
information.}
}
\value{
A \code{contact} S7 object describing who to contact for information about
the API, with fields \code{name}, \code{email}, and \code{url}.
}
\description{
A \code{contact} object provides contact information for the API.
}
\examples{
class_contact(
  "API Support",
  "support@example.com",
  "https://www.example.com/support"
)
}
\seealso{
info object:
\code{\link[=as_contact]{as_contact()}},
\code{\link[=as_info]{as_info()}},
\code{\link[=as_license]{as_license()}},
\code{\link[=as_origin]{as_origin()}},
\code{\link[=class_info]{class_info()}},
\code{\link[=class_license]{class_license()}},
\code{\link[=class_origin]{class_origin()}}
}
\concept{info}
