% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-license.R
\name{class_license}
\alias{class_license}
\title{License information for the API}
\usage{
class_license(
  name = character(),
  ...,
  identifier = character(),
  url = character()
)
}
\arguments{
\item{name}{Character scalar (optional). The license name used for the API.}

\item{...}{These dots are for future extensions and must be empty.}

\item{identifier}{Character scalar (optional). An
\href{https://spdx.org/spdx-specification-21-web-version#h.jxpfx0ykyb60}{SPDX}
license expression for the API. The \code{identifier} field is mutually
exclusive of the \code{url} field.}

\item{url}{Character scalar (optional). A URL to the license used for the
API. This must be in the form of a URL. The \code{url} field is mutually
exclusive of the \code{identifier} field.}
}
\value{
A \code{license} S7 object describing allowed usage of the API, with
fields \code{name}, \code{identifier}, and \code{url}.
}
\description{
A \code{license} object provides license information for the API to describe how
the API can be used.
}
\examples{
class_license(
  "Apache 2.0",
  identifier = "Apache-2.0"
)
class_license(
  "Apache 2.0",
  url = "https://opensource.org/license/apache-2-0/"
)
}
\seealso{
info object:
\code{\link[=as_contact]{as_contact()}},
\code{\link[=as_info]{as_info()}},
\code{\link[=as_license]{as_license()}},
\code{\link[=as_origin]{as_origin()}},
\code{\link[=class_contact]{class_contact()}},
\code{\link[=class_info]{class_info()}},
\code{\link[=class_origin]{class_origin()}}
}
\concept{info}
