% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/components-security_scheme-oauth2-implicit_flow.R
\name{class_oauth2_implicit_flow}
\alias{class_oauth2_implicit_flow}
\title{OAuth2 implicit flow object}
\usage{
class_oauth2_implicit_flow(
  authorization_url = character(),
  ...,
  refresh_url = character(),
  scopes = class_scopes()
)
}
\arguments{
\item{authorization_url}{Character vector (required). The authorization URL
to be used for this flow. This must be in the form of a URL. The OAuth2
standard requires the use of TLS.}

\item{...}{These dots are for future extensions and must be empty.}

\item{refresh_url}{Character scalar (optional). The URL to be used for
obtaining refresh tokens. This must be in the form of a URL. The OAuth2
standard requires the use of TLS.}

\item{scopes}{An optional \code{\link[=class_scopes]{class_scopes()}} object with the available scopes
for the OAuth2 security scheme.}
}
\value{
An \code{oauth2_implicit_flow} S7 object, with fields \code{authorization_url},
\code{refresh_url}, and \code{scopes}.
}
\description{
An \code{oauth2_implicit_flow} object describes the configuration for the OAuth
Implicit flow.
}
\examples{
class_oauth2_implicit_flow(
  authorization_url = "https://example.com/authorize",
  refresh_url = "https://example.com/refresh"
)
}
\seealso{
security scheme objects:
\code{\link[=as_api_key_security_scheme]{as_api_key_security_scheme()}},
\code{\link[=as_oauth2_authorization_code_flow]{as_oauth2_authorization_code_flow()}},
\code{\link[=as_oauth2_implicit_flow]{as_oauth2_implicit_flow()}},
\code{\link[=as_oauth2_security_scheme]{as_oauth2_security_scheme()}},
\code{\link[=as_oauth2_token_flow]{as_oauth2_token_flow()}},
\code{\link[=as_scopes]{as_scopes()}},
\code{\link[=as_security_scheme]{as_security_scheme()}},
\code{\link[=as_security_scheme_details]{as_security_scheme_details()}},
\code{\link[=as_security_schemes]{as_security_schemes()}},
\code{\link[=class_api_key_security_scheme]{class_api_key_security_scheme()}},
\code{\link[=class_oauth2_authorization_code_flow]{class_oauth2_authorization_code_flow()}},
\code{\link[=class_oauth2_security_scheme]{class_oauth2_security_scheme()}},
\code{\link[=class_oauth2_token_flow]{class_oauth2_token_flow()}},
\code{\link[=class_scopes]{class_scopes()}},
\code{\link[=class_security_scheme_details]{class_security_scheme_details()}},
\code{\link[=class_security_schemes]{class_security_schemes()}}

components object:
\code{\link[=as_api_key_security_scheme]{as_api_key_security_scheme()}},
\code{\link[=as_components]{as_components()}},
\code{\link[=as_oauth2_authorization_code_flow]{as_oauth2_authorization_code_flow()}},
\code{\link[=as_oauth2_implicit_flow]{as_oauth2_implicit_flow()}},
\code{\link[=as_oauth2_security_scheme]{as_oauth2_security_scheme()}},
\code{\link[=as_oauth2_token_flow]{as_oauth2_token_flow()}},
\code{\link[=as_reference]{as_reference()}},
\code{\link[=as_schema]{as_schema()}},
\code{\link[=as_scopes]{as_scopes()}},
\code{\link[=as_security_scheme]{as_security_scheme()}},
\code{\link[=as_security_scheme_details]{as_security_scheme_details()}},
\code{\link[=as_security_schemes]{as_security_schemes()}},
\code{\link[=class_api_key_security_scheme]{class_api_key_security_scheme()}},
\code{\link[=class_components]{class_components()}},
\code{\link[=class_oauth2_authorization_code_flow]{class_oauth2_authorization_code_flow()}},
\code{\link[=class_oauth2_security_scheme]{class_oauth2_security_scheme()}},
\code{\link[=class_oauth2_token_flow]{class_oauth2_token_flow()}},
\code{\link[=class_reference]{class_reference()}},
\code{\link[=class_schema]{class_schema()}},
\code{\link[=class_scopes]{class_scopes()}},
\code{\link[=class_security_scheme_details]{class_security_scheme_details()}},
\code{\link[=class_security_schemes]{class_security_schemes()}}
}
\concept{components}
\concept{components_security_schemes}
