% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-security_scheme-oauth2.R
\name{class_oauth2_security_scheme}
\alias{class_oauth2_security_scheme}
\title{OAuth2 security schemes}
\usage{
class_oauth2_security_scheme(
  ...,
  implicit_flow = class_oauth2_implicit_flow(),
  password_flow = class_oauth2_token_flow(),
  client_credentials_flow = class_oauth2_token_flow(),
  authorization_code_flow = class_oauth2_authorization_code_flow()
)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{implicit_flow}{An \code{oauth2_implicit_flow} object created with
\code{\link[=class_oauth2_implicit_flow]{class_oauth2_implicit_flow()}}.}

\item{password_flow, client_credentials_flow}{An \code{oauth2_token_flow} object
created with \code{\link[=class_oauth2_token_flow]{class_oauth2_token_flow()}}.}

\item{authorization_code_flow}{An \code{oauth2_authorization_code_flow} object
created with \code{\link[=class_oauth2_authorization_code_flow]{class_oauth2_authorization_code_flow()}}.}
}
\value{
An \code{oauth2_security_scheme} S7 object, with fields \code{implicit_flow},
\code{password_flow}, \code{client_credentials_flow}, and \code{authorization_code_flow}.
}
\description{
Defines an OAuth2 security scheme that can be used by the operations.
}
\examples{
class_oauth2_security_scheme()
class_oauth2_security_scheme(
  password_flow = class_oauth2_token_flow(token_url = "/tokens/passwords")
)
}
\seealso{
security scheme objects:
\code{\link[=as_api_key_security_scheme]{as_api_key_security_scheme()}},
\code{\link[=as_oauth2_authorization_code_flow]{as_oauth2_authorization_code_flow()}},
\code{\link[=as_oauth2_implicit_flow]{as_oauth2_implicit_flow()}},
\code{\link[=as_oauth2_security_scheme]{as_oauth2_security_scheme()}},
\code{\link[=as_oauth2_token_flow]{as_oauth2_token_flow()}},
\code{\link[=as_scopes]{as_scopes()}},
\code{\link[=as_security_scheme]{as_security_scheme()}},
\code{\link[=as_security_scheme_details]{as_security_scheme_details()}},
\code{\link[=as_security_schemes]{as_security_schemes()}},
\code{\link[=class_api_key_security_scheme]{class_api_key_security_scheme()}},
\code{\link[=class_oauth2_authorization_code_flow]{class_oauth2_authorization_code_flow()}},
\code{\link[=class_oauth2_implicit_flow]{class_oauth2_implicit_flow()}},
\code{\link[=class_oauth2_token_flow]{class_oauth2_token_flow()}},
\code{\link[=class_scopes]{class_scopes()}},
\code{\link[=class_security_scheme_details]{class_security_scheme_details()}},
\code{\link[=class_security_schemes]{class_security_schemes()}}

components object:
\code{\link[=as_api_key_security_scheme]{as_api_key_security_scheme()}},
\code{\link[=as_components]{as_components()}},
\code{\link[=as_oauth2_authorization_code_flow]{as_oauth2_authorization_code_flow()}},
\code{\link[=as_oauth2_implicit_flow]{as_oauth2_implicit_flow()}},
\code{\link[=as_oauth2_security_scheme]{as_oauth2_security_scheme()}},
\code{\link[=as_oauth2_token_flow]{as_oauth2_token_flow()}},
\code{\link[=as_reference]{as_reference()}},
\code{\link[=as_schema]{as_schema()}},
\code{\link[=as_scopes]{as_scopes()}},
\code{\link[=as_security_scheme]{as_security_scheme()}},
\code{\link[=as_security_scheme_details]{as_security_scheme_details()}},
\code{\link[=as_security_schemes]{as_security_schemes()}},
\code{\link[=class_api_key_security_scheme]{class_api_key_security_scheme()}},
\code{\link[=class_components]{class_components()}},
\code{\link[=class_oauth2_authorization_code_flow]{class_oauth2_authorization_code_flow()}},
\code{\link[=class_oauth2_implicit_flow]{class_oauth2_implicit_flow()}},
\code{\link[=class_oauth2_token_flow]{class_oauth2_token_flow()}},
\code{\link[=class_reference]{class_reference()}},
\code{\link[=class_schema]{class_schema()}},
\code{\link[=class_scopes]{class_scopes()}},
\code{\link[=class_security_scheme_details]{class_security_scheme_details()}},
\code{\link[=class_security_schemes]{class_security_schemes()}}
}
\concept{components}
\concept{components_security_schemes}
