% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-parameter.R
\name{class_parameter}
\alias{class_parameter}
\title{A single operation parameter}
\usage{
class_parameter(
  parameter_name = character(),
  schema = class_schema(),
  ...,
  location = c("query", "header", "path", "cookie"),
  style = parameter_style_default(location),
  description = character(),
  required = parameter_required_default(location),
  allow_empty_value = FALSE,
  allow_reserved = FALSE,
  explode = parameter_explode_default(style),
  deprecated = FALSE
)
}
\arguments{
\item{parameter_name}{Character scalar. The name of the parameter. Parameter
names are case sensitive.}

\item{schema}{\code{\link[=class_schema]{class_schema()}}. The schema defining the data type of the
parameter.}

\item{location}{Factor (or coercible to factor). The location of the
parameter in the API call. Possible values are "query", "header", "path",
or "cookie". Default "query".}

\item{style}{Factor (or coercible to factor, optional). Describes how the
parameter value will be serialized depending on the type of the parameter
value. Possible values are "matrix", "label", "form", "simple",
"spaceDelimited", "pipeDelimited", or "deepObject". Default values (set
automatically by \code{\link[=parameter_style_default]{parameter_style_default()}} based on the value of
\code{location}): for query - "form"; for path - "simple"; for header -
"simple"; for cookie - "form".}

\item{description}{Character scalar (optional). A brief description of the
parameter. This could contain examples of use. \href{https://spec.commonmark.org/}{CommonMark syntax} may be used for rich text
representation.}

\item{required}{Logical scalar (optional). Determines whether this parameter
is mandatory. If the parameter location is "path", this property is
required and its value must be \code{TRUE}. Otherwise, the property may be
included, and its default value is \code{FALSE}. Default values are set
automatically by \code{\link[=parameter_required_default]{parameter_required_default()}} based on the valu of
\code{location}.}

\item{allow_empty_value}{Logical scalar (optional). Sets the ability to pass
empty-valued parameters. This is valid only for "query"parameters, and
allows sending a parameter with an empty value. Default \code{FALSE}.}

\item{allow_reserved}{Logical scalar (optional). Determines whether the
parameter value should allow reserved characters, as defined by
\href{https://www.rfc-editor.org/rfc/rfc3986}{RFC3986} \verb{:/?#[]@!$&'()*+,;=} to
be included without percent-encoding. This property only applies to "query"
and "path" parameters.  Default \code{FALSE}.}

\item{explode}{Logical scalar (optional). When this is \code{TRUE}, parameter
values of type "array" or "object" generate separate parameters for each
value of the array or key-value pair of the map. For other types of
parameters this property has no effect. When \code{style} is "form", the default
value is \code{TRUE}. For all other \code{styles}, the default value is \code{FALSE}
(determined automatically via \code{\link[=parameter_explode_default]{parameter_explode_default()}}).}

\item{deprecated}{Logical scalar (optional). Specifies that a parameter is
deprecated and should be transitioned out of usage. Default \code{FALSE}.}
}
\value{
A \code{parameter} S7 object that describes a path or operation parameter.
}
\description{
A parameter for use in one or more \code{path} or \code{operation} objects.
}
\examples{
class_parameter()
class_parameter(
  parameter_name = "event_date",
  location = "query",
  schema = class_schema("string", format = "date-time"),
  required = TRUE
)
}
\seealso{
\code{\link[=as_parameter]{as_parameter()}} for coercing objects to \code{parameter}.
}
