% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-parameters.R
\name{class_parameters}
\alias{class_parameters}
\title{A collection of parameters and references to parameters}
\usage{
class_parameters(...)
}
\arguments{
\item{...}{One or more \code{\link[=abstract_security_scheme]{abstract_security_scheme()}} objects or a list of such
objects. These objects must be generated by
\code{\link[=class_api_key_security_scheme]{class_api_key_security_scheme()}} or \code{\link[=class_oauth2_security_scheme]{class_oauth2_security_scheme()}}
(\code{http_security_scheme()} and \code{open_id_connect_security_scheme()} coming
soon).}
}
\value{
A \code{security_scheme_details} object, which is a validated list of
\code{\link[=abstract_security_scheme]{abstract_security_scheme()}} objects.
}
\description{
The object provides a list of \code{\link[=class_parameter]{class_parameter()}} (or \code{\link[=class_reference]{class_reference()}}
pointing to \code{\link[=class_parameter]{class_parameter()}}).
}
