% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-rapid.R
\name{class_rapid}
\alias{class_rapid}
\title{R API description object}
\usage{
class_rapid(
  info = class_info(),
  ...,
  servers = class_servers(),
  components = class_components(),
  paths = class_paths(),
  security = class_security()
)
}
\arguments{
\item{info}{An \code{info} object defined by \code{\link[=class_info]{class_info()}}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{servers}{A \code{servers} object defined by \code{\link[=class_servers]{class_servers()}}.}

\item{components}{A \code{components} object defined by \code{\link[=class_components]{class_components()}}.}

\item{paths}{A \code{paths} object defined by \code{\link[=class_paths]{class_paths()}}.}

\item{security}{A \code{security} object defined by \code{\link[=class_security]{class_security()}}.}
}
\value{
A \code{rapid} S7 object, with properties \code{info}, \code{servers}, \code{components},
\code{paths}, and \code{security}.
}
\description{
An object that represents an API.
}
\examples{
class_rapid()
class_rapid(
  info = class_info(title = "A", version = "1"),
  servers = class_servers(
    url = "https://development.gigantic-server.com/v1"
  )
)
class_rapid(
  info = class_info(title = "A", version = "1"),
  servers = class_servers(
    url = c(
      "https://development.gigantic-server.com/v1",
      "https://staging.gigantic-server.com/v1",
      "https://api.gigantic-server.com/v1"
    ),
    description = c(
      "Development server",
      "Staging server",
      "Production server"
    )
  ),
  components = class_components(
    security_schemes = class_security_schemes(
      name = "a",
      details = class_security_scheme_details(
        class_api_key_security_scheme("parm", "query")
      )
    )
  )
)
}
\seealso{
rapid object:
\code{\link[=as_api_object]{as_api_object()}},
\code{\link[=as_rapid]{as_rapid()}},
\code{\link[=expand_servers]{expand_servers()}}
}
\concept{rapid}
