% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers-server_variables.R
\name{class_server_variables}
\alias{class_server_variables}
\title{A collection of string replacements for multiple servers}
\usage{
class_server_variables(...)
}
\arguments{
\item{...}{One or more \code{\link[=class_string_replacements]{class_string_replacements()}} objects, or a list of
\code{\link[=class_string_replacements]{class_string_replacements()}} objects.}
}
\value{
A \code{server_variables} S7 object, which is a validated list of
\code{\link[=class_string_replacements]{class_string_replacements()}} objects.
}
\description{
A list of string replacements objects, each of which is constructed with
\code{\link[=class_string_replacements]{class_string_replacements()}}.
}
\examples{
class_server_variables(
  list(class_string_replacements(), class_string_replacements())
)
}
\seealso{
servers object:
\code{\link[=as_server_variables]{as_server_variables()}},
\code{\link[=as_servers]{as_servers()}},
\code{\link[=as_string_replacements]{as_string_replacements()}},
\code{\link[=class_servers]{class_servers()}},
\code{\link[=class_string_replacements]{class_string_replacements()}}
}
\concept{servers}
