% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-01-contact.R
\name{contact}
\alias{contact}
\title{Contact information for the API}
\usage{
contact(name = class_missing, email = class_missing, url = class_missing)
}
\arguments{
\item{name}{Character scalar (optional). The identifying name of the contact
person/organization.}

\item{email}{Character scalar (optional). The email address of the contact
person/organization. This
\emph{must} be in the form of an email address.}

\item{url}{Character scalar (optional). The URL pointing to the contact
information.}
}
\value{
A \code{contact} S7 object, with fields \code{name}, \code{email}, and \code{url}.
}
\description{
Validate the contact information for an API.
}
\examples{
contact(
  "API Support",
  "support@example.com",
  "https://www.example.com/support"
)
}
