% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-conditions.R
\name{.pkg_inform}
\alias{.pkg_inform}
\title{Signal a package-scoped message}
\usage{
.pkg_inform(
  message,
  subclass,
  parent = NULL,
  call = caller_env(),
  message_env = caller_env(),
  ...
)
}
\arguments{
\item{message}{(\code{character}) The message for the new message condition.
Messages will be formatted with \code{\link[cli:cli_bullets]{cli::cli_bullets()}}.}

\item{subclass}{(\code{character}) Class(es) to assign to the message. Will be
prefixed by "\{package\}-message-".}

\item{parent}{A parent condition, as you might create during a
\code{\link[rlang:try_fetch]{rlang::try_fetch()}}. See \code{\link[rlang:abort]{rlang::abort()}} for additional information.}

\item{call}{(\code{environment}) The caller environment for error messages.}

\item{message_env}{(\code{environment}) The execution environment to use to
evaluate variables in error messages.}

\item{...}{Additional parameters passed to \code{\link[cli:cli_abort]{cli::cli_inform()}} and on to
\code{\link[rlang:abort]{rlang::inform()}}.}
}
\value{
\code{NULL}, invisibly (called for warning side effect).
}
\description{
Signal a package-scoped message
}
\keyword{internal}
