% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-shared-params.R
\name{.shared-parameters}
\alias{.shared-parameters}
\title{Parameters used in multiple functions}
\arguments{
\item{...}{Catch-all for unimplemented properties.}
}
\description{
Reused parameter definitions are gathered here for easier editing.
}
\keyword{internal}
