% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_sensitive_na.R
\name{handle_sensitive_na}
\alias{handle_sensitive_na}
\title{Handle Missing Values in the Sensitive Attribute}
\usage{
handle_sensitive_na(
  original_data,
  synthetic_data,
  sensitive_attribute,
  num_na_strategy = c("constant", "drop", "median"),
  num_constant_value = 0
)
}
\arguments{
\item{original_data}{The original dataset (data frame).}

\item{synthetic_data}{The synthetic dataset (data frame).}

\item{sensitive_attribute}{Name of the sensitive attribute (string).}

\item{num_na_strategy}{Strategy for handling missing numeric values: one of "drop", "constant", or "median".}

\item{num_constant_value}{Value used for imputation if \code{num_na_strategy = "constant"}.}
}
\value{
A list with cleaned \code{original_data} and \code{synthetic_data}.
}
\description{
Deals with missing values in the sensitive attribute for both original and synthetic data.
- For factors: replaces NAs with a new level "missing".
- For numerics: allows "drop", "constant", or "median" strategies.
}
