% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-01-license.R
\name{license}
\alias{license}
\title{License information for the API}
\usage{
license(name = character(), ..., identifier = character(), url = character())
}
\arguments{
\item{name}{Character scalar (optional). The license name used for the API.}

\item{...}{These dots are for future extensions and must be empty.}

\item{identifier}{Character scalar (optional). An
\href{https://spdx.org/spdx-specification-21-web-version#h.jxpfx0ykyb60}{SPDX}
license expression for the API. The \code{identifier} field is mutually
exclusive of the \code{url} field.}

\item{url}{Character scalar (optional). A URL to the license used for the
API. This \emph{must} be in the form of a URL. The \code{url} field is mutually
exclusive of the \code{identifier} field.}
}
\value{
A \code{license} S7 object, with fields \code{name}, \code{identifier}, and \code{url}.
}
\description{
Validate the license information for an API.
}
\examples{
license(
  "Apache 2.0",
  identifier = "Apache-2.0"
)
license(
  "Apache 2.0",
  url = "https://opensource.org/license/apache-2-0/"
)
}
\seealso{
\code{\link[=as_license]{as_license()}} for coercing objects to \code{license}.
}
