% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-security_scheme-oauth2-flows.R
\name{oauth2_flows}
\alias{oauth2_flows}
\title{OAuth2 flows object}
\usage{
oauth2_flows(refresh_url = class_missing, scopes = class_missing)
}
\arguments{
\item{refresh_url}{Character vector (optional). The URL to be used for
obtaining refresh tokens. This must be in the form of a URL. The OAuth2
standard requires the use of TLS.}

\item{scopes}{An optional \code{\link[=scopes_list]{scopes_list()}} object with the available scopes
for the OAuth2 security scheme.}
}
\description{
This is an abstract class that is used to define specific types of OAuth2
flows objects. Note: each OAuth2 flows object can hold multiple flows of that
type.
}
\seealso{
\code{\link[=ouath2_token_flows]{ouath2_token_flows()}}, \code{\link[=ouath2_implicit_flows]{ouath2_implicit_flows()}}, and
\code{\link[=ouath2_authorization_code_flows]{ouath2_authorization_code_flows()}}
}
