% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-security_scheme-oauth2.R
\name{oauth2_security_scheme}
\alias{oauth2_security_scheme}
\title{OAuth2 security schemes}
\usage{
oauth2_security_scheme(
  ...,
  implicit_flow = class_missing,
  password_flow = class_missing,
  client_credentials_flow = class_missing,
  authorization_code_flow = class_missing
)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{implicit_flow}{An \code{oauth2_implicit_flow} object created with
\code{\link[=oauth2_implicit_flow]{oauth2_implicit_flow()}}.}

\item{password_flow}{An \code{oauth2_token_flow} object created with
\code{\link[=oauth2_token_flow]{oauth2_token_flow()}}.}

\item{client_credentials_flow}{An \code{oauth2_token_flow} object created with
\code{\link[=oauth2_token_flow]{oauth2_token_flow()}}.}

\item{authorization_code_flow}{An \code{oauth2_authorization_code_flow} object
created with \code{\link[=oauth2_authorization_code_flow]{oauth2_authorization_code_flow()}}.}
}
\value{
An \code{oauth2_security_scheme} S7 object, with fields \code{implicit_flow},
\code{password_flow}, \code{client_credentials_flow}, and \code{authorization_code_flow}.
}
\description{
Defines one or more OAuth2 security schemes that can be used by the
operations.
}
\examples{
oauth2_security_scheme()
oauth2_security_scheme(
  password_flow = oauth2_token_flow(token_url = "/tokens/passwords")
)
}
