% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-security_scheme-oauth2-token_flows.R
\name{ouath2_token_flows}
\alias{ouath2_token_flows}
\title{OAuth2 token flows object}
\usage{
ouath2_token_flows(
  token_url = class_missing,
  ...,
  refresh_url = class_missing,
  scopes = class_missing
)
}
\arguments{
\item{token_url}{Character vector (required). The token URL to be used for
this flow. This must be in the form of a URL. The OAuth2 standard requires
the use of TLS.}

\item{...}{These dots are for future extensions and must be empty.}

\item{refresh_url}{Character vector (optional). The URL to be used for
obtaining refresh tokens. This must be in the form of a URL. The OAuth2
standard requires the use of TLS.}

\item{scopes}{An optional \code{\link[=scopes_list]{scopes_list()}} object with the available scopes
for the OAuth2 security scheme.}
}
\description{
An \code{ouath2_token_flows} object describes the configuration for the OAuth
Resource Owner Password flow or the OAuth Client Credentials flow (previously
called application in OpenAPI 2.0).
}
\examples{
ouath2_token_flows(
  token_url = "https://example.com/token",
  refresh_url = "https://example.com/refresh",
  scopes = scopes_list(
    scopes(
      name = c("server:read", "server:write"),
      description = c("Read server settings", "Write server settings")
    ),
    scopes(
      name = c("user:email_read", "user:email_write"),
      description = c("Read user email settings", "Write user email settings")
    )
  )
)
}
