% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-parameter.R
\name{parameter_explode_default}
\alias{parameter_explode_default}
\title{Choose the default explode value of a parameter}
\usage{
parameter_explode_default(style)
}
\arguments{
\item{style}{Factor (or coercible to factor, optional). Describes how the
parameter value will be serialized depending on the type of the parameter
value. Possible values are "matrix", "label", "form", "simple",
"spaceDelimited", "pipeDelimited", or "deepObject". Default values (set
automatically by \code{\link[=parameter_style_default]{parameter_style_default()}} based on the value of
\code{location}): for query - "form"; for path - "simple"; for header -
"simple"; for cookie - "form".}
}
\value{
\code{TRUE} for "form", \code{FALSE} for any other value, or a length-0 logical
vector if style is not defined.
}
\description{
This helper function chooses the default value for the \code{explode} property of
a \code{\link[=class_parameter]{class_parameter()}} based on the \code{style} property.
}
\examples{
parameter_explode_default()
parameter_explode_default("form")
parameter_explode_default("simple")
parameter_explode_default("spaceDelimited")
}
