% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-parameter.R
\name{parameter_required_default}
\alias{parameter_required_default}
\title{Choose the default required value of a parameter}
\usage{
parameter_required_default(location = c("query", "header", "path", "cookie"))
}
\arguments{
\item{location}{Factor (or coercible to factor). The location of the
parameter in the API call. Possible values are "query", "header", "path",
or "cookie". Default "query".}
}
\value{
\code{TRUE} for "path", \code{FALSE} for any other value, or a length-0 logical
vector if location is not defined.
}
\description{
This helper function chooses the default value for the \code{required} property of
a \code{\link[=class_parameter]{class_parameter()}} based on the \code{location} property.
}
\examples{
parameter_required_default()
parameter_required_default("query")
parameter_required_default("header")
parameter_required_default("path")
parameter_required_default("cookie")
}
