% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-parameter.R
\name{parameter_style_default}
\alias{parameter_style_default}
\title{Choose the default style of a parameter}
\usage{
parameter_style_default(location = c("query", "header", "path", "cookie"))
}
\arguments{
\item{location}{Factor (or coercible to factor). The location of the
parameter in the API call. Possible values are "query", "header", "path",
or "cookie".}
}
\value{
A character with one of "form" or "simple", or a length-0 character
vector.
}
\description{
This helper function chooses the default value for the \code{style} property of a
\code{\link[=class_parameter]{class_parameter()}} based on the \code{location} property.
}
\examples{
parameter_style_default()
parameter_style_default("query")
parameter_style_default("header")
parameter_style_default("path")
parameter_style_default("cookie")
}
