% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.rapid_result}
\alias{plot.rapid_result}
\title{################################################################################
Plot method for RAPID results}
\usage{
\method{plot}{rapid_result}(
  x,
  tau_range = seq(0, 1, by = 0.05),
  epsilon_range = seq(0, 20, by = 1),
  ...
)
}
\arguments{
\item{x}{A rapid_result object}

\item{tau_range}{Range of tau values for sensitivity curve (categorical)}

\item{epsilon_range}{Range of epsilon values for sensitivity curve (numeric)}

\item{...}{Additional arguments (unused)}
}
\description{
################################################################################
Plot method for RAPID results
}
