% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_sensitive.R
\name{predict_sensitive}
\alias{predict_sensitive}
\title{Predict Sensitive Attribute from Fitted Model}
\usage{
predict_sensitive(model_type, fit, original_data, sensitive_attribute)
}
\arguments{
\item{model_type}{A string: either "lm" or "rf".}

\item{fit}{The fitted model_type object.}

\item{original_data}{The original data (data frame) on which the predictions are made.}

\item{sensitive_attribute}{The name of the sensitive attribute (string).}
}
\value{
`B` — predicted values (vector for continuous, matrix or factor for categorical).
}
\description{
Predict Sensitive Attribute from Fitted Model
}
