% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.rapid_cv_result}
\alias{print.rapid_cv_result}
\title{Print method for RAPID cross-validation results}
\usage{
\method{print}{rapid_cv_result}(x, ...)
}
\arguments{
\item{x}{A rapid_cv_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for RAPID cross-validation results
}
