% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapid-package.R
\docType{package}
\name{rapid-package}
\alias{rapid}
\alias{rapid-package}
\title{rapid: R 'API' Descriptions}
\description{
Convert an 'API' description ('APID'), such as one that follows the 'OpenAPI Specification', to an R 'API' description object (a "rapid"). The rapid object follows the 'OpenAPI Specification' to make it easy to convert to and from 'API' documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rapid.api2r.org/}
  \item \url{https://github.com/api2r/rapid}
  \item Report bugs at \url{https://github.com/api2r/rapid/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

Authors:
\itemize{
  \item Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})
}

Other contributors:
\itemize{
  \item The Linux Foundation (OpenAPI Specification) [copyright holder]
}

}
\keyword{internal}
