% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_version.R
\name{rapid_spec_version}
\alias{rapid_spec_version}
\title{Construct the spec_version object for a rapid}
\usage{
rapid_spec_version(spec_type = NULL, spec_version = NULL, new_type = FALSE)
}
\arguments{
\item{spec_type}{A single string describing the type of spec, such as
"openapi", "swagger", or some other value.}

\item{spec_version}{A single character or other value indicating the version
number, such as "2.0" or "3.0.3" or "3". The value is validated for openapi
or swagger types, but left as-is for other types.}

\item{new_type}{A logical indicating whether this is a new type of spec
(something other than "openapi" or "swagger").}
}
\value{
An object consisting of a single version with a class specifying the
type of specification, or \code{NULL}.
}
\description{
A \code{spec_version} describes the specification schema from which the rapid was
derived.
}
\examples{
rapid_spec_version()
rapid_spec_version("openapi", 3)
rapid_spec_version("openapi", "3.0.3")
rapid_spec_version("swagger", 2)
rapid_spec_version("R-like", "funny-looking kid", new_type = TRUE)
}
