% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{caller_arg}
\alias{caller_env}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rlang}{\code{\link[rlang:caller_arg]{caller_arg()}}, \code{\link[rlang:caller_env]{caller_env()}}}
}}

