% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/components-01-security_scheme-02-oauth2-01-scopes.R
\name{scopes}
\alias{scopes}
\title{OAuth2 flow scopes objects}
\usage{
scopes(name = class_missing, description = class_missing)
}
\arguments{
\item{name}{Character (required). The name of the scope.}

\item{description}{Character (required). A short description of the scope.}
}
\value{
A \code{scopes} S7 object, with fields \code{name} and \code{description}.
}
\description{
The available scopes for an OAuth2 flow.
}
\examples{
scopes(
  name = c(
    "https://www.googleapis.com/auth/youtube",
    "https://www.googleapis.com/auth/youtube.channel-memberships.creator",
    "https://www.googleapis.com/auth/youtube.force-ssl",
    "https://www.googleapis.com/auth/youtube.readonly",
    "https://www.googleapis.com/auth/youtube.upload",
    "https://www.googleapis.com/auth/youtubepartner",
    "https://www.googleapis.com/auth/youtubepartner-channel-audit"
  ),
  description = c(
    "Manage your YouTube account",
    "See a list of your current active channel members, their current level, and when they became a member",
    "See, edit, and permanently delete your YouTube videos, ratings, comments and captions",
    "View your YouTube account",
    "Manage your YouTube videos",
    "View and manage your assets and associated content on YouTube",
    "View private information of your YouTube channel relevant during the audit process with a YouTube partner"
  )
)
}
\seealso{
\code{\link[=as_scopes]{as_scopes()}} for coercing objects to \code{scopes}.
}
