% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-security_scheme-oauth2-scopes_list.R
\name{scopes_list}
\alias{scopes_list}
\title{Multiple OAuth2 flow scopes objects}
\usage{
scopes_list(...)
}
\arguments{
\item{...}{One or more \code{\link[=scopes]{scopes()}} objects, or a list of \code{\link[=scopes]{scopes()}} objects.}
}
\value{
A \code{scopes_list} S7 object, which is a validated list of
\code{\link[=scopes]{scopes()}} objects.
}
\description{
The available scopes for multiple OAuth2 flows.
}
\examples{
scopes_list(
  scopes(
    name = c("server:read", "server:write"),
    description = c("Read server settings", "Write server settings")
  ),
  scopes(
    name = c("user:email_read", "user:email_write"),
    description = c("Read user email settings", "Write user email settings")
  )
)
}
\seealso{
\code{\link[=as_scopes_list]{as_scopes_list()}} for coercing objects to \code{scopes_list}.
}
