% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-security_scheme.R
\name{security_scheme}
\alias{security_scheme}
\title{Security scheme objects}
\usage{
security_scheme()
}
\description{
This is an abstract class that is used to define specific types of security
schemes.
}
\seealso{
\code{\link[=api_key_security_scheme]{api_key_security_scheme()}}, \code{\link[=oauth2_security_scheme]{oauth2_security_scheme()}}
}
\keyword{internal}
