% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-security_scheme_details.R
\name{security_scheme_details}
\alias{security_scheme_details}
\title{Details of API security schemes}
\usage{
security_scheme_details(...)
}
\arguments{
\item{...}{One or more \code{\link[=security_scheme]{security_scheme()}} objects or a list of such objects.
These objects must be generated by \code{\link[=api_key_security_scheme]{api_key_security_scheme()}} or
\code{\link[=oauth2_security_scheme]{oauth2_security_scheme()}} (\code{http_security_scheme()} and
\code{open_id_connect_security_scheme()} coming soon).}
}
\value{
A \code{\link[=security_scheme_details]{security_scheme_details()}} object, which is a validated list of
\code{\link[=security_scheme]{security_scheme()}} objects.
}
\description{
The object provides a list of details of security schemes for the API. Each
element within the list is a
}
\examples{
security_scheme_details(
  oauth2_security_scheme(
    password_flow = oauth2_token_flow(
      token_url = "/account/authorization",
      scopes = scopes(
        name = c("Catalog", "Commerce", "Playback", "Settings"),
        description = c(
          "Access all read-only content",
          "Perform account-level transactions",
          "Allow playback of restricted content",
          "Modify account settings"
        )
      )
    )
  ),
  oauth2_security_scheme(
    password_flow = oauth2_token_flow(
      token_url = "/account/profile/authorization",
      scopes = scopes(
        name = "Catalog",
        description = "Modify profile preferences and activity"
      )
    )
  ),
  api_key_security_scheme(
    parameter_name = "authorization",
    location = "header"
  ),
  api_key_security_scheme(
    parameter_name = "authorization",
    location = "header"
  )
)
}
