% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-security_scheme_type.R
\name{security_scheme_type}
\alias{security_scheme_type}
\title{Security scheme objects}
\usage{
security_scheme_type(name = class_missing, description = class_missing)
}
\arguments{
\item{name}{Character vector (required). The names of the members of this
scheme, used to identify valid schemes for a given operation.}

\item{description}{Character vector (optional). A description for each
security scheme. \href{https://spec.commonmark.org/}{CommonMark syntax} may be
used for rich text representation.}
}
\description{
This is an abstract class that is used to define specific types of security
schemes. Note: each security scheme object can hold multiple schemes of that
type.
}
\seealso{
\code{\link[=api_key_security_scheme]{api_key_security_scheme()}}
}
