% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-server.R
\name{server}
\alias{server}
\title{A collection of server variables for multiple servers}
\usage{
server(
  url = class_missing,
  description = class_missing,
  variables = class_missing
)
}
\arguments{
\item{url}{A list of \code{\link[=server_variable]{server_variable()}} objects.}

\item{description}{A list of \code{\link[=server_variable]{server_variable()}} objects.}

\item{variables}{\code{\link[=server_variable_list]{server_variable_list()}} object.}
}
\value{
A \code{server} S7 object, with properties \code{url}, \code{description}, and
\code{variables}.
}
\description{
Connectivity information for an API.
}
\examples{
server(
  url = c(
    "https://development.gigantic-server.com/v1",
    "https://staging.gigantic-server.com/v1",
    "https://api.gigantic-server.com/v1"
  ),
  description = c(
    "Development server",
    "Staging server",
    "Production server"
  )
)
server(
  url = "https://{username}.gigantic-server.com:{port}/{basePath}",
  description = "The production API server",
  variables = server_variable_list(server_variable(
    name = c("username", "port", "basePath"),
    default = c("demo", "8443", "v2"),
    description = c(
      "this value is assigned by the service provider, in this example `gigantic-server.com`",
      NA, NA
    ),
    enum = list(
      NULL,
      c("8443", "443"),
      NULL
    )
  ))
)
}
