% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers-02-server_variables.R
\name{server_variables}
\alias{server_variables}
\title{A collection of string replacements for multiple servers}
\usage{
server_variables(...)
}
\arguments{
\item{...}{One or more \code{\link[=string_replacements]{string_replacements()}} objects, or a list of
\code{\link[=string_replacements]{string_replacements()}} objects.}
}
\value{
A \code{server_variables} S7 object, which is a validated list of
\code{\link[=string_replacements]{string_replacements()}} objects.
}
\description{
A list of string replacements objects, each of which is constructed with
\code{\link[=string_replacements]{string_replacements()}}.
}
\examples{
server_variables(
  list(string_replacements(), string_replacements())
)
}
\seealso{
\code{\link[=as_server_variables]{as_server_variables()}} for coercing objects to \code{server_variables}.
}
