% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.rapid_result}
\alias{summary.rapid_result}
\title{################################################################################
Summary method for RAPID results}
\usage{
\method{summary}{rapid_result}(object, ...)
}
\arguments{
\item{object}{A rapid_result object}

\item{...}{Additional arguments (unused)}
}
\description{
################################################################################
Summary method for RAPID results
}
