% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers-02-variables.R
\name{variables}
\alias{variables}
\title{A collection of server variables for multiple servers}
\usage{
variables(...)
}
\arguments{
\item{...}{One or more \code{\link[=server_variable]{server_variable()}} objects, or a list of
\code{\link[=server_variable]{server_variable()}} objects.}
}
\value{
A \code{variables} S7 object, which is a validated list of
\code{\link[=server_variable]{server_variable()}} objects.
}
\description{
A list of server variable objects, each of which is constructed with
\code{\link[=server_variable]{server_variable()}}.
}
\examples{
variables(
  list(server_variable(), server_variable())
)
}
