% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-contact.R
\name{api_contact}
\alias{api_contact}
\title{Contact information for the API}
\usage{
api_contact(name = NULL, url = NULL, email = NULL)
}
\arguments{
\item{name}{The identifying name of the contact person/organization.}

\item{url}{The URL pointing to the contact information. This \emph{must} be in the
form of a URL.}

\item{email}{The email address of the contact person/organization. This
\emph{must} be in the form of an email address.}
}
\value{
A \code{rapid_contact} object: a list with fields \code{name}, \code{url}, and
\code{email}.
}
\description{
Validate the contact information for an API.
}
\examples{
api_contact(
  "API Support",
  "https://www.example.com/support",
  "support@example.com"
)
}
