% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-security_scheme-api_key.R
\name{api_key_security_scheme}
\alias{api_key_security_scheme}
\title{API key security schemes}
\usage{
api_key_security_scheme(
  name = class_missing,
  parameter_name = class_missing,
  `in` = class_missing,
  ...,
  description = class_missing
)
}
\arguments{
\item{name}{Character (required). The names of the members of this scheme,
used to identify valid schemes for a given operation.}

\item{parameter_name}{Character (required). The names of the header, query or
cookie parameters to be used.}

\item{description}{Character (optional). A description for each security
scheme. \href{https://spec.commonmark.org/}{CommonMark syntax} \emph{may} be used for
rich text representation.}

\item{`in`}{Character (required). The location of the API key. Valid values
are "query", "header" or "cookie".}
}
\value{
An \code{api_key_security_scheme} S7 object, with fields \code{name},
\code{description}, \code{parameter_name}, and \verb{in}.
}
\description{
Defines one or more API key security schemes that can be used by the
operations.
}
\examples{
api_key_security_scheme(
  "api_key_example",
  description = "API key security scheme example",
  parameter_name = "Authorization",
  `in` = "header"
)
}
\seealso{
\code{\link[=as_api_key_security_scheme]{as_api_key_security_scheme()}} for coercing objects to
\code{api_key_security_scheme}.
}
