% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-security_scheme-api_key.R
\name{as_api_key_security_scheme}
\alias{as_api_key_security_scheme}
\title{Coerce lists and character vectors to API key security schemes}
\usage{
as_api_key_security_scheme(x, ...)
}
\arguments{
\item{x}{The object to coerce. Must be empty or be a list or character vector
with names "name" and either "in" or "location", or names that can be
coerced to those names via \code{\link[snakecase:caseconverter]{snakecase::to_snake_case()}}. Additional names
are ignored.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
An \code{api_key_security_scheme} as returned by
\code{\link[=api_key_security_scheme]{api_key_security_scheme()}}.
}
\description{
\code{as_api_key_security_scheme()} turns an existing object into an
\code{api_key_security_scheme}. This is in contrast with
\code{\link[=api_key_security_scheme]{api_key_security_scheme()}}, which builds an \code{api_key_security_scheme} from
individual properties.
}
