% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-security_scheme-api_key.R
\name{as_api_key_security_scheme}
\alias{as_api_key_security_scheme}
\title{Coerce lists to API key security schemes}
\usage{
as_api_key_security_scheme(x, ...)
}
\arguments{
\item{x}{The object to coerce. Must be empty or be a list of named lists,
each with names "description", "name", and/or "in". Additional names are
ignored.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A \code{api_key_security_scheme} as returned by
\code{\link[=api_key_security_scheme]{api_key_security_scheme()}}.
}
\description{
\code{as_api_key_security_scheme()} turns an existing object into a
\code{api_key_security_scheme}. This is in contrast with
\code{\link[=api_key_security_scheme]{api_key_security_scheme()}}, which builds a \code{api_key_security_scheme} from
individual properties.
}
