% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as_api_object}
\alias{as_api_object}
\title{Convert to a rapid-style object}
\usage{
as_api_object(
  x,
  target_class,
  ...,
  alternate_names = NULL,
  arg = caller_arg(x),
  call = caller_env()
)
}
\arguments{
\item{x}{The object to coerce. Must be empty, or be a named list or character
vector having names corresponding to the parameter of the \code{target_class},
or names that can be coerced to those names via
\code{\link[snakecase:caseconverter]{snakecase::to_snake_case()}}. Extra names are ignored.}

\item{target_class}{The S7 class to which the object should be converted.}

\item{...}{These dots are for future extensions and must be empty.}

\item{alternate_names}{Character vector (optional). An optional named
character vector, where the names are the names as they might appear in
\code{x}, and the values are the corresponding properties.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
An object with the specified \code{target_class}.
}
\description{
Convert an object into an object with a rapid-style class.
}
