% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{as_component_collection}
\alias{as_component_collection}
\title{Coerce lists to component_collection objects}
\usage{
as_component_collection(x, ..., arg = caller_arg(x), call = caller_env())
}
\arguments{
\item{x}{The object to coerce. Must be empty or be a list containing a single
list named "security_schemes", or a name that can be coerced to
"security_schemes" via \code{\link[snakecase:caseconverter]{snakecase::to_snake_case()}}. Additional names are
ignored.}

\item{...}{These dots are for future extensions and must be empty.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A \code{component_collection} object as returned by
\code{\link[=component_collection]{component_collection()}}.
}
\description{
\code{as_component_collection()} turns an existing object into a
\code{component_collection} object. This is in contrast with
\code{\link[=component_collection]{component_collection()}}, which builds a \code{component_collection} from
individual properties.
}
\examples{
as_component_collection()
as_component_collection(list(
  securitySchemes = list(
    accountAuth = list(
      description = "Account JWT token",
      flows = list(
        password = list(
          scopes = list(
            Catalog = "Access all read-only content",
            Commerce = "Perform account-level transactions",
            Playback = "Allow playback of restricted content",
            Settings = "Modify account settings"
          ),
          tokenUrl = "/account/authorization"
        )
      ),
      type = "oauth2"
    ),
    profileAuth = list(
      description = "Profile JWT token",
      flows = list(
        password = list(
          scopes = list(
            Catalog = "Modify profile preferences and activity"
          ),
          tokenUrl = "/account/profile/authorization"
        )
      ),
      type = "oauth2"
    ),
    resetPasswordAuth = list(
      `in` = "header",
      name = "authorization",
      type = "apiKey"
    ),
    verifyEmailAuth = list(
      `in` = "header",
      name = "authorization",
      type = "apiKey"
    )
  )
))
}
