% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-contact.R
\name{as_contact}
\alias{as_contact}
\title{Coerce lists and character vectors to contacts}
\usage{
as_contact(x, ...)
}
\arguments{
\item{x}{The object to coerce. Must be empty or have names "name", "email",
and/or "url", or names that can be coerced to those names via
\code{\link[snakecase:caseconverter]{snakecase::to_snake_case()}}. Extra names are ignored. This object should
describe a single point of contact.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A \code{contact} as returned by \code{\link[=contact]{contact()}}.
}
\description{
\code{as_contact()} turns an existing object into a \code{contact}. This is in contrast
with \code{\link[=contact]{contact()}}, which builds a \code{contact} from individual properties.
}
\examples{
as_contact()
as_contact(list(name = "Jon Harmon", email = "jonthegeek@gmail.com"))
}
