% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{as_info}
\alias{as_info}
\title{Coerce lists and character vectors to info objects}
\usage{
as_info(x, ..., arg = caller_arg(x), call = caller_env())
}
\arguments{
\item{x}{The object to coerce. Must be empty or have names "title",
"version", "contact", "description", "license", "summary",
"terms_of_service", and/or "origin" (or "x-origin", which will be coerced
to "origin"), or names that can be coerced to those names via
\code{\link[snakecase:caseconverter]{snakecase::to_snake_case()}}. Extra names are ignored. This object should
describe a single API.}

\item{...}{These dots are for future extensions and must be empty.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
An \code{info} object as returned by \code{\link[=class_info]{class_info()}}.
}
\description{
\code{as_info()} turns an existing object into an \code{info} object. This is in
contrast with \code{\link[=class_info]{class_info()}}, which builds an \code{info} from individual properties.
}
\examples{
as_info()
as_info(list(title = "My Cool API", version = "1.0.0"))
}
