% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-zz-info.R
\name{as_info}
\alias{as_info}
\title{Coerce lists and character vectors to info objects}
\usage{
as_info(x, ...)
}
\arguments{
\item{x}{The object to coerce. Must be empty or have names "title",
"version", "contact", "description", "license", "summary", and/or
"terms_of_service". Extra names are ignored.}
}
\value{
An \code{info} object as returned by \code{\link[=info]{info()}}.
}
\description{
\code{as_info()} turns an existing object into an \code{info} object. This is in
contrast with \code{\link[=info]{info()}}, which builds an \code{info} from individual properties.
}
\examples{
as_info()
as_info(list(title = "My Cool API", version = "1.0.0"))
}
