% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-license.R
\name{as_license}
\alias{as_license}
\title{Coerce lists and character vectors to licenses}
\usage{
as_license(x, ..., arg = caller_arg(x), call = caller_env())
}
\arguments{
\item{x}{The object to coerce. Must be empty or have names "name",
"identifier", and/or "url", or names that can be coerced to those names via
\code{\link[snakecase:caseconverter]{snakecase::to_snake_case()}}. Extra names are ignored. This object should
describe a single license.}

\item{...}{These dots are for future extensions and must be empty.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A \code{license} as returned by \code{\link[=class_license]{class_license()}}.
}
\description{
\code{as_license()} turns an existing object into a \code{license}. This is in contrast
with \code{\link[=class_license]{class_license()}}, which builds a \code{license} from individual properties.
}
\examples{
as_license()
as_license(list(name = "Apache 2.0", identifier = "Apache-2.0"))
}
