% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/components-security_scheme-oauth2-authorization_code_flow.R
\name{as_oauth2_authorization_code_flow}
\alias{as_oauth2_authorization_code_flow}
\title{Coerce lists and character vectors to OAuth2 authorization code flows}
\usage{
as_oauth2_authorization_code_flow(
  x,
  ...,
  arg = caller_arg(x),
  call = caller_env()
)
}
\arguments{
\item{x}{The object to coerce. Must be empty or be a list of named lists,
each with names "refresh_url", "scopes", "authorization_url", and/or
"token_url", or names that can be coerced to those names via
\code{\link[snakecase:caseconverter]{snakecase::to_snake_case()}}. Additional names are ignored.}

\item{...}{These dots are for future extensions and must be empty.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
An \code{oauth2_authorization_code_flow} as returned by
\code{\link[=oauth2_authorization_code_flow]{oauth2_authorization_code_flow()}}.
}
\description{
\code{as_oauth2_authorization_code_flow()} turns an existing object into an
\code{oauth2_authorization_code_flow}. This is in contrast with
\code{\link[=oauth2_authorization_code_flow]{oauth2_authorization_code_flow()}}, which builds an
\code{oauth2_authorization_code_flow} from individual properties.
}
