% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-security_scheme-oauth2.R
\name{as_oauth2_security_scheme}
\alias{as_oauth2_security_scheme}
\title{Coerce lists to oauth2 security schemes}
\usage{
as_oauth2_security_scheme(x, ...)
}
\arguments{
\item{x}{The object to coerce. Must be empty or be a list with an object
named "flows" or a name that can be coerced to "flows" via
\code{\link[snakecase:caseconverter]{snakecase::to_snake_case()}}. Additional names are ignored.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
An \code{oauth2_security_scheme} as returned by
\code{\link[=oauth2_security_scheme]{oauth2_security_scheme()}}.
}
\description{
\code{as_oauth2_security_scheme()} turns an existing object into an
\code{oauth2_security_scheme}. This is in contrast with
\code{\link[=oauth2_security_scheme]{oauth2_security_scheme()}}, which builds an \code{oauth2_security_scheme} from
individual properties.
}
