% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-security_scheme-oauth2-token_flow.R
\name{as_oauth2_token_flow}
\alias{as_oauth2_token_flow}
\title{Coerce lists to OAuth2 token flow}
\usage{
as_oauth2_token_flow(x, ...)
}
\arguments{
\item{x}{The object to coerce. Must be empty or be a list of named lists,
each with names "refresh_url", "scopes", and/or "token_url". Additional
names are ignored.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A \code{oauth2_token_flow} as returned by \code{\link[=oauth2_token_flow]{oauth2_token_flow()}}.
}
\description{
\code{as_oauth2_token_flow()} turns an existing object into a
\code{oauth2_token_flow}. This is in contrast with \code{\link[=oauth2_token_flow]{oauth2_token_flow()}}, which
builds a \code{oauth2_token_flow} from individual properties.
}
