% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-origin.R
\name{as_origin}
\alias{as_origin}
\title{Coerce lists and character vectors to class_origin}
\usage{
as_origin(x, ...)
}
\arguments{
\item{x}{The object to coerce. Must be empty or have names "url", "format",
and/or "version", or names that can be coerced to those names via
\code{\link[snakecase:caseconverter]{snakecase::to_snake_case()}}. Extra names are ignored. This object should
describe a single origin for this API description.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A \code{class_origin} as returned by \code{\link[=class_origin]{class_origin()}}.
}
\description{
\code{as_origin()} turns an existing object into a \code{class_origin}. This is in
contrast with \code{\link[=class_origin]{class_origin()}}, which builds a \code{class_origin} from individual
properties.
}
\examples{
as_origin()
as_origin(
  list(
    list(
      format = "openapi",
      url = "https://api.open.fec.gov/swagger/",
      version = "3.0"
    )
  )
)
}
