% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{as_paths}
\alias{as_paths}
\title{Coerce objects to paths}
\usage{
as_paths(x, ...)
}
\arguments{
\item{x}{The object to coerce. Must be empty or be a \code{data.frame()}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A \code{paths} object as returned by \code{\link[=class_paths]{class_paths()}}.
}
\description{
\code{as_paths()} turns an existing object into a \code{paths} object. This is in
contrast with \code{\link[=class_paths]{class_paths()}}, which builds a \code{paths} object from individual
properties. In practice, \code{\link[=class_paths]{class_paths()}} and \code{as_paths()} are currently
functionally identical. However, in the future, \code{as_paths()} will coerce
other valid objects to the expected shape.
}
\examples{
as_paths()
as_paths(mtcars)
}
\seealso{
paths object
\code{\link{class_paths}()}
}
\concept{paths}
