% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-reference.R
\name{as_reference}
\alias{as_reference}
\title{Coerce lists and character vectors to references}
\usage{
as_reference(x, ..., arg = caller_arg(x), call = caller_env())
}
\arguments{
\item{x}{The object to coerce. Must be empty or have names "type",
"nullable", "description", and/or "format", or names that can be coerced to
those names via \code{\link[snakecase:caseconverter]{snakecase::to_snake_case()}}. Extra names are ignored. This
object should describe a single reference.}

\item{...}{These dots are for future extensions and must be empty.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A \code{reference} as returned by \code{\link[=class_reference]{class_reference()}}.
}
\description{
\code{as_reference()} turns an existing object into a \code{reference}. This is in contrast
with \code{\link[=class_reference]{class_reference()}}, which builds a \code{reference} from individual properties.
}
\examples{
as_reference()
as_reference(list(`$ref` = "#/components/schemas/Pet"))
}
