% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-security_scheme-oauth2-scopes.R
\name{as_scopes}
\alias{as_scopes}
\title{Coerce lists and character vectors to scopes}
\usage{
as_scopes(x, ...)
}
\arguments{
\item{x}{The object to coerce. Must be coercible to a named character vector.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A \code{scopes} as returned by \code{\link[=scopes]{scopes()}}.
}
\description{
\code{as_scopes()} turns an existing object into a
\code{scopes}. This is in contrast with \code{\link[=scopes]{scopes()}},
which builds a \code{scopes} from individual properties.
}
