% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security.R
\name{as_security}
\alias{as_security}
\title{Coerce lists to as_security objects}
\usage{
as_security(x, ...)
}
\arguments{
\item{x}{The object to coerce. Must be empty or be a list containing a single
list named "security_schemes", or a name that can be coerced to
"security_schemes" via \code{\link[snakecase:caseconverter]{snakecase::to_snake_case()}}. Additional names are
ignored.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A \code{security} object as returned by \code{\link[=class_security]{class_security()}}.
}
\description{
\code{as_security()} turns an existing object into a \code{security} object. This is in
contrast with \code{\link[=class_security]{class_security()}}, which builds a \code{security} from individual
properties.
}
\examples{
as_security()
as_security(
  list(
    list(
      oauth2 = c("user", "user:email", "user:follow")
    ),
    list(internalApiKey = list())
  )
)
}
\seealso{
security object
\code{\link{class_security}()}
}
\concept{security}
