% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-security_scheme.R
\name{as_security_scheme}
\alias{as_security_scheme}
\title{Coerce lists to security_scheme objects}
\usage{
as_security_scheme(x, ...)
}
\arguments{
\item{x}{The object to coerce. Must be empty or be a named list, with at
least an element \code{type}. The \code{type} element is processed through
\code{\link[snakecase:caseconverter]{snakecase::to_snake_case()}}, and then must be one of "api_key", "oauth2",
or "oauth_2".}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A \code{security_scheme} object as returned by \code{\link[=api_key_security_scheme]{api_key_security_scheme()}}
or \code{\link[=oauth2_security_scheme]{oauth2_security_scheme()}}.
}
\description{
\code{as_security_scheme()} turns an existing object into a \code{security_scheme}
object. It uses the \code{type} element of such objects to determine which type of
security scheme to construct.
}
\examples{
as_security_scheme()
as_security_scheme(
  list(
    description = "Account JWT token",
    flows = list(
      password = list(
        scopes = list(
          Catalog = "Access all read-only content",
          Commerce = "Perform account-level transactions",
          Playback = "Allow playback of restricted content",
          Settings = "Modify account settings"
        ),
        tokenUrl = "/account/authorization"
      )
    ),
    type = "oauth2"
  )
)
as_security_scheme(
  list(
    description = "Profile JWT token",
    flows = list(
      password = list(
        scopes = list(
          Catalog = "Modify profile preferences and activity (bookmarks, watch list)"
        ),
        tokenUrl = "/account/profile/authorization"
      )
    ),
    type = "oauth2"
  )
)
as_security_scheme(
  list(
    `in` = "header",
    name = "authorization",
    type = "apiKey"
  )
)
}
