% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers-02-server_variables.R
\name{as_server_variables}
\alias{as_server_variables}
\title{Coerce lists and character vectors to server_variables}
\usage{
as_server_variables(x, ...)
}
\arguments{
\item{x}{The object to coerce. Must be empty or be a list of objects that can
be coerced to \code{string_replacements} objects via \code{\link[=as_string_replacements]{as_string_replacements()}}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A \code{server_variables} object as returned by \code{\link[=server_variables]{server_variables()}}.
}
\description{
\code{as_server_variables()} turns an existing object into a \code{server_variables}
object. This is in contrast with \code{\link[=server_variables]{server_variables()}}, which builds a
\code{server_variables} object from individual properties.
}
\examples{
as_server_variables()
as_server_variables(
  list(
    list(
      username = c(default = "demo", description = "Name of the user.")
    ),
    list(
      username = c(
        default = "demo",
        description = "Name of the user."
      ),
      port = list(
        default = "8443",
        enum = c("8443", "443")
      )
    )
  )
)
}
