% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers-zz-servers.R
\name{as_servers}
\alias{as_servers}
\title{Coerce lists and character vectors to servers}
\usage{
as_servers(x, ...)
}
\arguments{
\item{x}{The object to coerce. Must be empty or have names "name", "email",
and/or "url". Extra names are ignored.}
}
\value{
A \code{servers} as returned by \code{\link[=servers]{servers()}}.
}
\description{
\code{as_servers()} turns an existing object into a \code{servers}. This is in
contrast with \code{\link[=servers]{servers()}}, which builds a \code{servers} from individual
properties.
}
\examples{
as_servers()
as_servers(list(name = "Jon Harmon", email = "jonthegeek@gmail.com"))
}
