% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-contact.R
\name{class_contact}
\alias{class_contact}
\title{Contact information for the API}
\usage{
class_contact(name = class_missing, email = class_missing, url = class_missing)
}
\arguments{
\item{name}{Character scalar (optional). The identifying name of the contact
person or organization.}

\item{email}{Character scalar (optional). The email address of the contact
person/organization. This must be in the form of an email address.}

\item{url}{Character scalar (optional). The URL pointing to the contact
information.}
}
\value{
A \code{contact} S7 object describing who to contact for information about
the API, with fields \code{name}, \code{email}, and \code{url}.
}
\description{
A \code{contact} object provides contact information for the API.
}
\examples{
class_contact(
  "API Support",
  "support@example.com",
  "https://www.example.com/support"
)
}
\seealso{
info object
\code{\link{as_contact}()},
\code{\link{as_info}()},
\code{\link{as_license}()},
\code{\link{as_origin}()},
\code{\link{class_info}()},
\code{\link{class_license}()},
\code{\link{class_origin}()}
}
\concept{info}
