% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{class_info}
\alias{class_info}
\title{Information about the API}
\usage{
class_info(
  title = character(),
  version = character(),
  ...,
  contact = class_contact(),
  description = character(),
  license = class_license(),
  summary = character(),
  terms_of_service = character(),
  origin = class_origin()
)
}
\arguments{
\item{title}{Character scalar. The title of the API. Required when the object
is not empty.}

\item{version}{Character scalar. The version of the API description (which is
distinct from the OpenAPI Specification version or the API implementation
version).  Required when the object is not empty.}

\item{...}{These dots are for future extensions and must be empty.}

\item{contact}{The contact information for the exposed API, generated via
\code{\link[=class_contact]{class_contact()}}.}

\item{description}{Character scalar (optional). A description of the API.
\href{https://spec.commonmark.org/}{CommonMark syntax} may be used for rich text
representation.}

\item{license}{The license information for the exposed API, generated via
\code{\link[=license]{license()}}.}

\item{summary}{Character scalar (optional). A short summary of the API.}

\item{terms_of_service}{Character scalar (optional). A URL to the Terms of
Service for the API.}

\item{origin}{The url and related information about the document used to
build the API description. This is used to resolve relative paths in the
API description. Note: This is not part of the OpenAPI Specification, but
is sometimes supplied as "x-origin".}
}
\value{
An \code{info} S7 object with metadata describing a single API.
}
\description{
The object provides metadata about the API. The metadata may be used by the
clients if needed, and may be presented in editing or documentation
generation tools for convenience.
}
\examples{
class_info()
class_info(
  title = "My Cool API",
  version = "1.0.2",
  license = class_license(
    name = "Apache 2.0",
    url = "https://opensource.org/license/apache-2-0/"
  )
)
class_info(
  title = "My Abbreviated API",
  version = "2.0.0",
  origin = "https://root.url"
)
}
\seealso{
info object
\code{\link{as_contact}()},
\code{\link{as_info}()},
\code{\link{as_license}()},
\code{\link{as_origin}()},
\code{\link{class_contact}()},
\code{\link{class_license}()},
\code{\link{class_origin}()}
}
\concept{info}
