% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/components-security_scheme-oauth2-authorization_code_flow.R
\name{class_oauth2_authorization_code_flow}
\alias{class_oauth2_authorization_code_flow}
\title{OAuth2 authorization code flow object}
\usage{
class_oauth2_authorization_code_flow(
  authorization_url = character(),
  token_url = character(),
  ...,
  refresh_url = character(),
  scopes = class_scopes()
)
}
\arguments{
\item{authorization_url}{Character vector (required). The authorization URL
to be used for this flow. This must be in the form of a URL. The OAuth2
standard requires the use of TLS.}

\item{token_url}{Character vector (required). The token URL to be used for
this flow. This must be in the form of a URL. The OAuth2 standard requires
the use of TLS.}

\item{...}{These dots are for future extensions and must be empty.}

\item{refresh_url}{Character scalar (optional). The URL to be used for
obtaining refresh tokens. This must be in the form of a URL. The OAuth2
standard requires the use of TLS.}

\item{scopes}{An optional \code{\link[=class_scopes]{class_scopes()}} object with the available scopes
for the OAuth2 security scheme.}
}
\description{
An \code{oauth2_authorization_code_flow} object describes the configuration for
the OAuth Authorization Code flow. Previously called \code{accessCode} in OpenAPI
2.0.
}
\examples{
class_oauth2_authorization_code_flow(
  authorization_url = "https://example.com/authorize",
  token_url = "https://example.com/token",
  refresh_url = "https://example.com/refresh",
  scopes = class_scopes(
    name = c("server:read", "server:write"),
    description = c("Read server settings", "Write server settings")
  )
)
}
