% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-security_scheme-oauth2.R
\name{class_oauth2_security_scheme}
\alias{class_oauth2_security_scheme}
\title{OAuth2 security schemes}
\usage{
class_oauth2_security_scheme(
  ...,
  implicit_flow = class_oauth2_implicit_flow(),
  password_flow = class_oauth2_token_flow(),
  client_credentials_flow = class_oauth2_token_flow(),
  authorization_code_flow = class_oauth2_authorization_code_flow()
)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{implicit_flow}{An \code{oauth2_implicit_flow} object created with
\code{\link[=class_oauth2_implicit_flow]{class_oauth2_implicit_flow()}}.}

\item{password_flow, client_credentials_flow}{An \code{oauth2_token_flow} object created with
\code{\link[=class_oauth2_token_flow]{class_oauth2_token_flow()}}.}

\item{authorization_code_flow}{An \code{oauth2_authorization_code_flow} object
created with \code{\link[=class_oauth2_authorization_code_flow]{class_oauth2_authorization_code_flow()}}.}
}
\value{
An \code{oauth2_security_scheme} S7 object, with fields \code{implicit_flow},
\code{password_flow}, \code{client_credentials_flow}, and \code{authorization_code_flow}.
}
\description{
Defines an OAuth2 security scheme that can be used by the operations.
}
\examples{
class_oauth2_security_scheme()
class_oauth2_security_scheme(
  password_flow = class_oauth2_token_flow(token_url = "/tokens/passwords")
)
}
