% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-origin.R
\name{class_origin}
\alias{class_origin}
\title{Source information for the API description}
\usage{
class_origin(
  url = character(),
  ...,
  format = character(),
  version = character()
)
}
\arguments{
\item{url}{Character (required). The URL(s) where the document(s) can be
found.}

\item{...}{These dots are for future extensions and must be empty.}

\item{format}{Character scalar (optional). The format of the document.
Presently this will likely always be "openapi".}

\item{version}{Character scalar (optional). The specification version
(relative to the \code{format}) used in defining the document. Not to be
confused with the version of the API description itself. Most often this
will be "3.0" (as in "OpenAPI Specification version 3.0"), "3.1", or a
patch version of those.}
}
\value{
A \code{class_origin} S7 object describing where to find the API
description, with fields \code{url}, \code{format}, and \code{version}.
}
\description{
A \code{class_origin} object provides information about the primary source
document(s) used to build an API.
}
\examples{
class_origin(
  "https://api.open.fec.gov/swagger/",
  format = "openapi",
  version = "3.0"
)
}
\seealso{
\code{\link[=as_origin]{as_origin()}} for coercing objects to \code{class_origin}.
}
