% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-reference.R
\name{class_reference}
\alias{class_reference}
\title{A simple object for referencing other components in the API description}
\usage{
class_reference(
  ref_uri = class_missing,
  summary = class_missing,
  description = class_missing
)
}
\arguments{
\item{ref_uri}{Character scalar. The reference identifier. This must be in
the form of a URI.}

\item{summary}{Character scalar (optional). A short summary which by default
should override that of the referenced component. If the referenced
object-type does not allow a summary field, then this field has no effect.}

\item{description}{Character scalar (optional). A description which by
default should override that of the referenced component. \href{https://spec.commonmark.org/}{CommonMark syntax} may be used for rich text
representation. If the referenced object-type does not allow a description
field, then this field has no effect.}
}
\value{
A \code{reference} S7 object pointing (by name) to another part of the
\code{rapid} object.
}
\description{
The \code{reference} object allows for reuse of components between different parts
of the API description. These objects are currently simple character
references, but may change structure in the future to automatically resolve
references.
}
\examples{
class_reference("#/components/schemas/Pet")
}
\seealso{
\code{\link[=as_reference]{as_reference()}} for coercing objects to \code{reference}.
}
