% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security.R
\name{class_security}
\alias{class_security}
\title{Security schemes required to execute an operation}
\usage{
class_security(name = character(), ..., required_scopes = list())
}
\arguments{
\item{name}{Character vector (required). The names must correspond to
security schemes declared in the \code{security_schemes} property of a
\code{\link[=class_components]{class_components()}}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{required_scopes}{A list of character vectors, each of which describe
the scopes required for this security scheme. The vector corresponding to a
given \code{name} can be empty.}
}
\value{
A \code{security} S7 object with references of security required for
operations.
}
\description{
The object lists the required security schemes to execute an operation or
operations.
}
\examples{
class_security()
class_security(
  name = c("oauth2", "internalApiKey"),
  required_scopes = list(
    c("user", "user:email", "user:follow"),
    character()
  )
)
}
\seealso{
security object
\code{\link{as_security}()}
}
\concept{security}
