% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-security_scheme_details.R
\name{class_security_scheme_details}
\alias{class_security_scheme_details}
\title{Details of API security schemes}
\usage{
class_security_scheme_details(...)
}
\arguments{
\item{...}{One or more \code{\link[=abstract_security_scheme]{abstract_security_scheme()}} objects or a list of such
objects. These objects must be generated by
\code{\link[=class_api_key_security_scheme]{class_api_key_security_scheme()}} or \code{\link[=class_oauth2_security_scheme]{class_oauth2_security_scheme()}}
(\code{http_security_scheme()} and \code{open_id_connect_security_scheme()} coming
soon).}
}
\value{
A \code{security_scheme_details} object, which is a validated list of
\code{\link[=abstract_security_scheme]{abstract_security_scheme()}} objects.
}
\description{
The object provides a list of details of security schemes for the API. Each
element within the list is an \code{\link[=abstract_security_scheme]{abstract_security_scheme()}} object.
}
\examples{
class_security_scheme_details(
  class_oauth2_security_scheme(
    password_flow = class_oauth2_token_flow(
      token_url = "/account/authorization",
      scopes = class_scopes(
        name = c("Catalog", "Commerce", "Playback", "Settings"),
        description = c(
          "Access all read-only content",
          "Perform account-level transactions",
          "Allow playback of restricted content",
          "Modify account settings"
        )
      )
    )
  ),
  class_oauth2_security_scheme(
    password_flow = class_oauth2_token_flow(
      token_url = "/account/profile/authorization",
      scopes = class_scopes(
        name = "Catalog",
        description = "Modify profile preferences and activity"
      )
    )
  ),
  class_api_key_security_scheme(
    parameter_name = "authorization",
    location = "header"
  ),
  class_api_key_security_scheme(
    parameter_name = "authorization",
    location = "header"
  )
)
}
