% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-security_schemes.R
\name{class_security_schemes}
\alias{class_security_schemes}
\title{Reusable security schemes for an API}
\usage{
class_security_schemes(
  name = character(),
  details = class_security_scheme_details(),
  ...,
  description = character()
)
}
\arguments{
\item{name}{Character vector (required). Names by which security schemes will
be referenced.}

\item{details}{The details of each security scheme, as a
\code{\link[=class_security_scheme_details]{class_security_scheme_details()}} object.}

\item{...}{These dots are for future extensions and must be empty.}

\item{description}{Character vector (optional). A short description for the
security schemes. \href{https://spec.commonmark.org/}{CommonMark syntax} may be
used for rich text representation.}
}
\value{
A \code{security_schemes} S7 object with details about security available
for operations.
}
\description{
The object provides reusable security schemes for the API. These schemes may
be referenced by name in the top-level \code{security} object or within \code{paths}
objects.
}
\examples{
class_security_schemes()
class_security_schemes(
  name = c(
    "accountAuth",
    "resetPasswordAuth"
  ),
  details = class_security_scheme_details(
    class_oauth2_security_scheme(
      password_flow = class_oauth2_token_flow(
        token_url = "/account/authorization",
        scopes = class_scopes(
          name = c("Catalog", "Commerce", "Playback", "Settings"),
          description = c(
            "Access all read-only content",
            "Perform account-level transactions",
            "Allow playback of restricted content",
            "Modify account settings"
          )
        )
      )
    ),
    class_api_key_security_scheme(
      parameter_name = "authorization",
      location = "header"
    )
  )
)
}
\seealso{
security scheme objects
\code{\link{as_api_key_security_scheme}()},
\code{\link{as_oauth2_authorization_code_flow}()},
\code{\link{as_oauth2_implicit_flow}()},
\code{\link{as_oauth2_security_scheme}()},
\code{\link{as_oauth2_token_flow}()},
\code{\link{as_scopes}()},
\code{\link{as_security_scheme}()},
\code{\link{as_security_scheme_details}()},
\code{\link{as_security_schemes}()},
\code{\link{class_api_key_security_scheme}()},
\code{\link{class_oauth2_authorization_code_flow}()},
\code{\link{class_oauth2_implicit_flow}()},
\code{\link{class_oauth2_security_scheme}()},
\code{\link{class_oauth2_token_flow}()},
\code{\link{class_scopes}()},
\code{\link{class_security_scheme_details}()}

components object
\code{\link{as_api_key_security_scheme}()},
\code{\link{as_components}()},
\code{\link{as_oauth2_authorization_code_flow}()},
\code{\link{as_oauth2_implicit_flow}()},
\code{\link{as_oauth2_security_scheme}()},
\code{\link{as_oauth2_token_flow}()},
\code{\link{as_reference}()},
\code{\link{as_schema}()},
\code{\link{as_scopes}()},
\code{\link{as_security_scheme}()},
\code{\link{as_security_scheme_details}()},
\code{\link{as_security_schemes}()},
\code{\link{class_api_key_security_scheme}()},
\code{\link{class_components}()},
\code{\link{class_oauth2_authorization_code_flow}()},
\code{\link{class_oauth2_implicit_flow}()},
\code{\link{class_oauth2_security_scheme}()},
\code{\link{class_oauth2_token_flow}()},
\code{\link{class_reference}()},
\code{\link{class_schema}()},
\code{\link{class_scopes}()},
\code{\link{class_security_scheme_details}()}
}
\concept{components}
\concept{components_security_schemes}
