% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers-string_replacements.R
\name{class_string_replacements}
\alias{class_string_replacements}
\title{A set of variables for server URL template substitution}
\usage{
class_string_replacements(
  name = character(),
  default = character(),
  ...,
  enum = list(),
  description = character()
)
}
\arguments{
\item{name}{Character vector (required). The names of the variables.}

\item{default}{Character vector (required). The default value to use for
substitution of each variable, which shall be sent if an alternate value is
not supplied. Note this behavior is different than the Schema Object’s
treatment of default values, because in those cases parameter values are
optional. If the \code{enum} is defined, the value must exist in the enum’s
values.}

\item{...}{These dots are for future extensions and must be empty.}

\item{enum}{List of (potentially empty) character vectors (optional). An
enumeration of valid string values to be used if the substitution options
are from a limited set.}

\item{description}{Character vector (optional). An optional description for
each server variable. \href{https://spec.commonmark.org/}{CommonMark syntax} may
be used for rich text representation.}
}
\value{
A \code{string_replacements} S7 object describing a collection of string
variables for a single server, with fields \code{name}, \code{default}, \code{enum}, and
\code{description}.
}
\description{
A \code{string_replacements} object describes server variable properties used for
substitution in a single server’s URL template.
}
\examples{
class_string_replacements(
  "username",
  "demo",
  enum = c("demo", "other"),
  description = "The active user's folder."
)
}
\seealso{
servers object
\code{\link{as_server_variables}()},
\code{\link{as_servers}()},
\code{\link{as_string_replacements}()},
\code{\link{class_server_variables}()},
\code{\link{class_servers}()}
}
\concept{servers}
