% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolute_paths.R
\name{expand_servers}
\alias{expand_servers}
\title{Expand server urls to absolute paths}
\usage{
expand_servers(x, ...)
}
\arguments{
\item{x}{The object to update. Must be a \code{rapid}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A \code{rapid} object as returned by \code{\link[=class_rapid]{class_rapid()}}, with absolute server
paths.
}
\description{
\code{expand_servers()} uses the \code{origin} property of a \code{rapid} object to expand
the \code{servers} \code{url} property to an absolute path.
}
