% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-zz-info.R
\name{info}
\alias{info}
\title{Information about the API}
\usage{
info(
  title = class_missing,
  version = class_missing,
  contact = class_missing,
  description = class_missing,
  license = class_missing,
  summary = class_missing,
  terms_of_service = class_missing
)
}
\arguments{
\item{title}{The title of the API. Required when the object is not empty.}

\item{version}{The version of the API document (which is distinct from the
OpenAPI Specification version or the API implementation version).  Required
when the object is not empty.}

\item{contact}{The contact information for the exposed API, generated via
\code{\link[=contact]{contact()}}.}

\item{description}{A description of the API. \href{https://spec.commonmark.org/}{CommonMark syntax} \emph{may} be used for rich text
representation.}

\item{license}{The license information for the exposed API, generated via
\code{\link[=license]{license()}}.}

\item{summary}{A short summary of the API.}

\item{terms_of_service}{A URL to the Terms of Service for the API.}
}
\value{
An \code{info} S7 object.
}
\description{
The object provides metadata about the API. The metadata \emph{may} be used by the
clients if needed, and \emph{may} be presented in editing or documentation
generation tools for convenience.
}
\examples{
info()
info(
  title = "My Cool API",
  version = "1.0.2",
  license = license(
    name = "Apache 2.0",
    url = "https://opensource.org/license/apache-2-0/"
  )
)
}
