% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapid-package.R
\docType{package}
\name{rapid-package}
\alias{rapid}
\alias{rapid-package}
\title{rapid: R 'API' Definitions}
\description{
Convert an 'API' document ('APID'), such as one that follows the 'OpenAPI Specification', to an R 'API' definition object.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jonthegeek.github.io/rapid/}
  \item \url{https://github.com/jonthegeek/rapid}
  \item Report bugs at \url{https://github.com/jonthegeek/rapid/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
