% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-rapid.R
\name{rapid}
\alias{rapid}
\title{R API description object}
\usage{
rapid(
  info = class_missing,
  ...,
  servers = class_missing,
  components = component_collection()
)
}
\arguments{
\item{info}{An \code{info} object defined by \code{\link[=info]{info()}}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{servers}{A \code{servers} object defined by \code{\link[=servers]{servers()}}.}

\item{components}{A \code{component_collection} object defined by
\code{\link[=component_collection]{component_collection()}}.}
}
\value{
A \code{rapid} S7 object, with properties \code{info}, \code{servers}, and
\code{components}.
}
\description{
An object that represents an API.
}
\examples{
rapid()
rapid(
  info = info(title = "A", version = "1"),
  servers = servers(
    url = "https://development.gigantic-server.com/v1"
  )
)
rapid(
  info = info(title = "A", version = "1"),
  servers = servers(
    url = c(
      "https://development.gigantic-server.com/v1",
      "https://staging.gigantic-server.com/v1",
      "https://api.gigantic-server.com/v1"
    ),
    description = c(
      "Development server",
      "Staging server",
      "Production server"
    )
  ),
  components = component_collection(
    security_schemes = security_scheme_collection(
      name = "a",
      details = security_scheme_details(
        api_key_security_scheme("parm", "query")
      )
    )
  )
)
}
\seealso{
\code{\link[=as_rapid]{as_rapid()}} for coercing objects to \code{rapid}.
}
