% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-rapid.R
\name{rapid}
\alias{rapid}
\title{R API definition object}
\usage{
rapid(info = class_missing, servers = class_missing)
}
\arguments{
\item{info}{An \code{api_info} object defined by \code{\link[=api_info]{api_info()}}.}

\item{servers}{A \code{servers} object defined by \code{\link[=servers]{servers()}}.}
}
\value{
A \code{rapid} S7 object, with properties \code{info} and \code{servers}.
}
\description{
An object that represents an API.
}
\examples{
rapid()
rapid(
  info = api_info(title = "A", version = "1"),
  servers(
    url = "https://development.gigantic-server.com/v1"
  )
)
rapid(
  info = api_info(title = "A", version = "1"),
  servers(
    url = c(
      "https://development.gigantic-server.com/v1",
      "https://staging.gigantic-server.com/v1",
      "https://api.gigantic-server.com/v1"
    ),
    description = c(
      "Development server",
      "Staging server",
      "Production server"
    )
  )
)
}
