% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-01-security_scheme-01-type.R
\name{security_scheme}
\alias{security_scheme}
\title{Security schemes}
\usage{
security_scheme(name = class_missing, description = class_missing)
}
\arguments{
\item{name}{Character (required). The name of this scheme, used to identify
valid schemes for a given operation.}

\item{description}{Character (optional). A description for the security
scheme. \href{https://spec.commonmark.org/}{CommonMark syntax} \emph{may} be used for
rich text representation.}
}
\description{
This is an abstract class that is only used to define specific types of
security schemes.
}
\seealso{
\code{\link[=api_key_security_scheme]{api_key_security_scheme()}}
}
