% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers-02-server_variable_list.R
\name{server_variable_list}
\alias{server_variable_list}
\title{A collection of server variables for multiple servers}
\usage{
server_variable_list(..., apid_list = NULL)
}
\arguments{
\item{...}{One or more \code{\link[=server_variable]{server_variable()}} objects, or a list of
\code{\link[=server_variable]{server_variable()}} objects.}

\item{apid_list}{An API document as a list, such as that obtained by reading
an OAS document with \code{\link[yaml:read_yaml]{yaml::read_yaml()}}.}
}
\value{
A \code{server_variable_list} S7 object, which is a validated list of
\code{\link[=server_variable]{server_variable()}} objects.
}
\description{
A list of server variable objects, each of which is constructed with
\code{\link[=server_variable]{server_variable()}}.
}
\examples{
server_variable_list(
  list(server_variable(), server_variable())
)
}
