% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers-zz-servers.R
\name{servers}
\alias{servers}
\title{An object representing a collection of servers}
\usage{
servers(
  url = class_missing,
  description = class_missing,
  variables = class_missing
)
}
\arguments{
\item{url}{Character vector (required). The urls of the target hosts. These
urls support \code{\link[=string_replacements]{string_replacements()}}. Variable substitutions will be made
when a variable is named in \{brackets\}.}

\item{description}{Character vector (optional). Strings describing the hosts
designated by \code{url}. \href{https://spec.commonmark.org/}{CommonMark syntax}
\emph{may} be used for rich text representation.}

\item{variables}{A \code{\link[=server_variables]{server_variables()}} object.}
}
\value{
A \code{servers} S7 object, with properties \code{url}, \code{description}, and
\code{variables}.
}
\description{
Connectivity information for an API.
}
\examples{
servers(
  url = c(
    "https://development.gigantic-server.com/v1",
    "https://staging.gigantic-server.com/v1",
    "https://api.gigantic-server.com/v1"
  ),
  description = c(
    "Development server",
    "Staging server",
    "Production server"
  )
)
servers(
  url = "https://{username}.gigantic-server.com:{port}/{basePath}",
  description = "The production API server",
  variables = server_variables(string_replacements(
    name = c("username", "port", "basePath"),
    default = c("demo", "8443", "v2"),
    description = c(
      "The active user's folder.",
      NA, NA
    ),
    enum = list(
      NULL,
      c("8443", "443"),
      NULL
    )
  ))
)
}
\seealso{
\code{\link[=as_servers]{as_servers()}} for coercing objects to \code{servers}.
}
