% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers-zz-servers.R
\name{servers}
\alias{servers}
\title{A collection of server variables for multiple servers}
\usage{
servers(
  url = class_missing,
  description = class_missing,
  variables = class_missing
)
}
\arguments{
\item{url}{A character vector of urls.}

\item{description}{A character vector of (usually brief) descriptions of
those urls.}

\item{variables}{A \code{\link[=server_variable_list]{server_variable_list()}} object.}
}
\value{
A \code{servers} S7 object, with properties \code{url}, \code{description}, and
\code{variables}.
}
\description{
Connectivity information for an API.
}
\examples{
servers(
  url = c(
    "https://development.gigantic-server.com/v1",
    "https://staging.gigantic-server.com/v1",
    "https://api.gigantic-server.com/v1"
  ),
  description = c(
    "Development server",
    "Staging server",
    "Production server"
  )
)
servers(
  url = "https://{username}.gigantic-server.com:{port}/{basePath}",
  description = "The production API server",
  variables = server_variable_list(server_variable(
    name = c("username", "port", "basePath"),
    default = c("demo", "8443", "v2"),
    description = c(
      "The active user's folder.",
      NA, NA
    ),
    enum = list(
      NULL,
      c("8443", "443"),
      NULL
    )
  ))
)
}
