% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers-zz-servers.R
\name{servers}
\alias{servers}
\title{A collection of server variables for multiple servers}
\usage{
servers(
  url = S7::class_missing,
  description = S7::class_missing,
  variables = S7::class_missing,
  ...,
  apid_list = NULL
)
}
\arguments{
\item{url}{A list of \code{\link[=server_variable]{server_variable()}} objects.}

\item{description}{A list of \code{\link[=server_variable]{server_variable()}} objects.}

\item{variables}{\code{\link[=server_variable_list]{server_variable_list()}} object.}

\item{...}{Catch-all for unimplemented properties.}

\item{apid_list}{An API document as a list, such as that obtained by reading
an OAS document with \code{\link[yaml:read_yaml]{yaml::read_yaml()}}.}
}
\value{
A \code{servers} S7 object, with properties \code{url}, \code{description}, and
\code{variables}.
}
\description{
Connectivity information for an API.
}
\examples{
servers(
  url = c(
    "https://development.gigantic-server.com/v1",
    "https://staging.gigantic-server.com/v1",
    "https://api.gigantic-server.com/v1"
  ),
  description = c(
    "Development server",
    "Staging server",
    "Production server"
  )
)
servers(
  url = "https://{username}.gigantic-server.com:{port}/{basePath}",
  description = "The production API server",
  variables = server_variable_list(server_variable(
    name = c("username", "port", "basePath"),
    default = c("demo", "8443", "v2"),
    description = c(
      "The active user's folder.",
      NA, NA
    ),
    enum = list(
      NULL,
      c("8443", "443"),
      NULL
    )
  ))
)
}
